% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_keys.R
\name{add_key}
\alias{add_key}
\title{Internal assignment of value to a JoinKeys object}
\usage{
add_key(join_keys_obj, dataset_1, dataset_2 = dataset_1, value)
}
\arguments{
\item{join_keys_obj}{(\code{JoinKeys}) object to extract the join keys}

\item{dataset_1}{(\code{character}) name of first dataset.}

\item{dataset_2}{(\code{character}) name of second dataset.}

\item{value}{(\code{character} vector) value to assign.}
}
\description{
Internal assignment of value to a JoinKeys object
}
\keyword{internal}
