% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_keys-append.R
\name{append_join_keys}
\alias{append_join_keys}
\title{Append elements to join_keys}
\usage{
append_join_keys(x, values, after = length(join_keys(x)))
}
\arguments{
\item{x}{A teal_data object}

\item{values}{(\code{join_keys}) object to append}

\item{after}{Integer, the position after which the elements are to be appended.
If negative or zero, the values are prepended to the join_keys.
If missing, the values are appended at the end.}
}
\value{
A teal_data object with updated join_keys
}
\description{
This function appends elements to the join_keys slot of a teal_data object.
It follows the same pattern as base R's append() function.
}
\examples{
data <- teal_data()
keys1 <- join_keys(
  join_key("ADSL", "ADSL", "USUBJID"),
  join_key("ADAE", "ADAE", "USUBJID"),
  join_key("ADSL", "ADAE", c(USUBJID = "USUBJID"))
)
data <- append_join_keys(data, keys1)
join_keys(data)data

}
