% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_cdisc.R
\name{as_cdisc}
\alias{as_cdisc}
\alias{as_cdisc.TealDataset}
\alias{as_cdisc.TealDatasetConnector}
\title{Convert a \code{TealDataset(Connector)} object to a \code{CDISCTealDataset(Connector)} object}
\usage{
as_cdisc(
  x,
  parent = if (identical(get_dataname(x), "ADSL")) character(0) else "ADSL"
)

\method{as_cdisc}{TealDataset}(
  x,
  parent = if (identical(get_dataname(x), "ADSL")) character(0) else "ADSL"
)

\method{as_cdisc}{TealDatasetConnector}(
  x,
  parent = if (identical(get_dataname(x), "ADSL")) character(0) else "ADSL"
)
}
\arguments{
\item{x}{an object of \code{TealDataset} or \code{TealDatasetConnector} class}

\item{parent}{(\code{character}, optional) parent dataset name}
}
\value{
(\code{CDISCTealDataset} or \code{CDISCTealDatasetConnector}) object
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
\details{
Convert a \code{TealDataset(Connector)} object to a \code{CDISCTealDataset(Connector)} object
}
\note{
If passed a \code{CDISC}-flavored object it returns the unmodified object.
}
\examples{
# TealDataset --------

as_cdisc(
  dataset(
    "ADSL",
    example_cdisc_data("ADSL"),
    keys = get_cdisc_keys("ADSL"),
    code = "ADSL <- teal.data::rADSL"
  )
)
as_cdisc(
  dataset(
    "ADAE",
    example_cdisc_data("ADAE"),
    keys = get_cdisc_keys("ADAE"),
    code = "ADAE <- teal.data::rADAE"
  ),
  parent = "ADSL"
)
# TealDatasetConnector --------

pull_fun_adsl <- callable_function(
  function() {
    example_cdisc_data("ADSL")
  }
)
as_cdisc(
  dataset_connector(
    "ADSL",
    pull_fun_adsl,
    keys = get_cdisc_keys("ADSL")
  )
)

pull_fun_adae <- callable_function(
  function() {
    example_cdisc_data("ADAE")
  }
)
as_cdisc(
  dataset_connector(
    "ADAE",
    pull_fun_adae,
    keys = get_cdisc_keys("ADAE")
  ),
  parent = "ADSL"
)
}
