% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_join_keys.R
\name{assert_compatible_keys}
\alias{assert_compatible_keys}
\title{Helper function to assert if two key sets contain incompatible keys}
\usage{
assert_compatible_keys(join_key_1, join_key_2)
}
\description{
return TRUE if compatible, throw error otherwise
}
\keyword{internal}
