% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_keys.R
\name{assert_join_keys}
\alias{assert_join_keys}
\alias{test_join_keys}
\alias{expect_join_keys}
\title{Check the JoinKeys class membership of an argument}
\usage{
assert_join_keys(x, .var.name = checkmate::vname(x))

test_join_keys(x)

expect_join_keys(x)
}
\arguments{
\item{x}{[any]\cr
Object to check.}

\item{.var.name}{[\code{character(1)}]\cr
Name of the checked object to print in assertions. Defaults to
the heuristic implemented in \code{\link[checkmate]{vname}}.}

\item{extra_classes}{(\code{character} vector) with extra classes to check. Can be used}
}
\value{
\code{x} invisibly
}
\description{
Check the JoinKeys class membership of an argument
}
\keyword{internal}
