% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_keys.R
\name{assert_join_keys_alike}
\alias{assert_join_keys_alike}
\alias{check_join_keys_alike}
\title{Assert the \code{join_keys} class membership of an argument}
\usage{
assert_join_keys_alike(x, .var.name = checkmate::vname(x), add = NULL)

check_join_keys_alike(x)
}
\arguments{
\item{x}{[any]\cr
Object to check.}

\item{.var.name}{[\code{character(1)}]\cr
Name of the checked object to print in assertions. Defaults to
the heuristic implemented in \code{\link[checkmate]{vname}}.}

\item{add}{[\code{AssertCollection}]\cr
Collection to store assertion messages. See \code{\link[checkmate]{AssertCollection}}.}
}
\value{
\code{x} invisibly
}
\description{
Assert the \code{join_keys} class membership of an argument
}
\keyword{internal}
