% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_keys.R
\name{assert_non_conflicting_keys}
\alias{assert_non_conflicting_keys}
\title{Helper function to assert non-conflicting keys}
\usage{
assert_non_conflicting_keys(.x, .y)
}
\value{
TRUE if compatible, throw error otherwise
}
\description{
Function asserts that there are no key sets which refer to the same pair of
datasets.
}
\keyword{internal}
