% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_keys.R
\name{assert_parent_child}
\alias{assert_parent_child}
\title{Helper function checks the parent-child relations are valid}
\usage{
assert_parent_child(join_keys_obj)
}
\arguments{
\item{join_keys_obj}{(\code{JoinKeys}) object to assert validity of relations}
}
\value{
\code{join_keys_obj} invisibly
}
\description{
Helper function checks the parent-child relations are valid
}
\keyword{internal}
