% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_cdisc_join_keys.R
\name{build_cdisc_join_keys}
\alias{build_cdisc_join_keys}
\title{Helper method to build \code{default_cdisc_join_keys}}
\usage{
build_cdisc_join_keys(default_cdisc_keys)
}
\arguments{
\item{default_cdisc_keys}{(\code{list}) default definition of primary and foreign
keys for CDISC datasets}
}
\description{
Helper method to build \code{default_cdisc_join_keys}
}
\keyword{internal}
