% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CallablePythonCode.R
\name{CallablePythonCode}
\alias{CallablePythonCode}
\title{A \code{CallablePythonCode} class of objects}
\description{
A \code{CallablePythonCode} class of objects

A \code{CallablePythonCode} class of objects
}
\keyword{internal}
\section{Super classes}{
\code{\link[teal.data:Callable]{teal.data::Callable}} -> \code{\link[teal.data:CallableFunction]{teal.data::CallableFunction}} -> \code{CallablePythonCode}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CallablePythonCode-new}{\code{CallablePythonCode$new()}}
\item \href{#method-CallablePythonCode-set_object}{\code{CallablePythonCode$set_object()}}
\item \href{#method-CallablePythonCode-run}{\code{CallablePythonCode$run()}}
\item \href{#method-CallablePythonCode-clone}{\code{CallablePythonCode$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="teal.data" data-topic="Callable" data-id="assign_to_env"><a href='../../teal.data/html/Callable.html#method-Callable-assign_to_env'><code>teal.data::Callable$assign_to_env()</code></a></span></li>
<li><span class="pkg-link" data-pkg="teal.data" data-topic="Callable" data-id="get_error_message"><a href='../../teal.data/html/Callable.html#method-Callable-get_error_message'><code>teal.data::Callable$get_error_message()</code></a></span></li>
<li><span class="pkg-link" data-pkg="teal.data" data-topic="Callable" data-id="is_failed"><a href='../../teal.data/html/Callable.html#method-Callable-is_failed'><code>teal.data::Callable$is_failed()</code></a></span></li>
<li><span class="pkg-link" data-pkg="teal.data" data-topic="CallableFunction" data-id="get_args"><a href='../../teal.data/html/CallableFunction.html#method-CallableFunction-get_args'><code>teal.data::CallableFunction$get_args()</code></a></span></li>
<li><span class="pkg-link" data-pkg="teal.data" data-topic="CallableFunction" data-id="get_call"><a href='../../teal.data/html/CallableFunction.html#method-CallableFunction-get_call'><code>teal.data::CallableFunction$get_call()</code></a></span></li>
<li><span class="pkg-link" data-pkg="teal.data" data-topic="CallableFunction" data-id="set_arg_value"><a href='../../teal.data/html/CallableFunction.html#method-CallableFunction-set_arg_value'><code>teal.data::CallableFunction$set_arg_value()</code></a></span></li>
<li><span class="pkg-link" data-pkg="teal.data" data-topic="CallableFunction" data-id="set_args"><a href='../../teal.data/html/CallableFunction.html#method-CallableFunction-set_args'><code>teal.data::CallableFunction$set_args()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CallablePythonCode-new"></a>}}
\if{latex}{\out{\hypertarget{method-CallablePythonCode-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{CallablePythonCode} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CallablePythonCode$new(fun, env = new.env(parent = parent.env(globalenv())))}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{fun}}{(\code{function})\cr
function to be evaluated in class. Function should be named}

\item{\code{env}}{(\code{environment})\cr
environment where the result of python code evaluation are stored}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
new \code{CallablePythonCode} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CallablePythonCode-set_object"></a>}}
\if{latex}{\out{\hypertarget{method-CallablePythonCode-set_object}{}}}
\subsection{Method \code{set_object()}}{
For scripts and code that contain multiple objects, save the name
of the object that corresponds to the final dataset of interest.
This is required for running python scripts with \code{reticulate}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CallablePythonCode$set_object(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{(\code{character}) the name of the object produced by the code
or script.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
(\code{self}) invisibly for chaining.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CallablePythonCode-run"></a>}}
\if{latex}{\out{\hypertarget{method-CallablePythonCode-run}{}}}
\subsection{Method \code{run()}}{
Execute \code{Callable} python code.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CallablePythonCode$run(args = NULL, try = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{args}}{(\code{NULL} or named \code{list})\cr
supplied for callable functions only, these are dynamic arguments passed to
\code{reticulate::py_run_string} or \code{reticulate::py_run_file}. Dynamic arguments
are executed in this call and are not saved which means that \code{self$get_call()}
won't include them later.}

\item{\code{try}}{(\code{logical} value)\cr
whether perform function evaluation inside \code{try} clause}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
nothing or output from function depending on \code{return}
argument. If \code{run} fails it will return object of class \code{simple-error} error
when \code{try = TRUE} or will stop if \code{try = FALSE}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CallablePythonCode-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CallablePythonCode-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CallablePythonCode$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
