% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdisc_data.R
\name{cdisc_data}
\alias{cdisc_data}
\title{Data input for teal app}
\usage{
cdisc_data(..., join_keys = teal.data::join_keys(), code = "", check = FALSE)
}
\arguments{
\item{...}{(\code{TealDataConnector}, \code{TealDataset}, \code{TealDatasetConnector}, \code{any})\cr
\itemize{
\item When one of the \verb{Teal*} objects are provided, then function returns \code{TealData} object.
This way of specifying data is deprecated and will be removed in the next release.
\item From version 0.4.0, one can provide any object as a named argument and function will
return \code{teal_data} object. Objects provided in \code{...} will be stored in \code{teal_data} environment
under the same name as the argument name.
}}

\item{join_keys}{(\code{JoinKeys}) or a single (\code{JoinKeySet})\cr
(optional) object with datasets column names used for joining.
If empty then it would be automatically derived basing on intersection of datasets primary keys.
For ADAM datasets it would be automatically derived.}

\item{code}{(\code{character}) code to reproduce the datasets.}

\item{check}{(\code{logical}) reproducibility check - whether to perform a check that the pre-processing
code included in the object definitions actually produces those objects.
If \code{check} is true and preprocessing code is empty an error will be thrown.}
}
\value{
a \code{TealData} or \code{teal_data} object
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
Function is a wrapper around \code{\link[=teal_data]{teal_data()}} and guesses \code{join_keys}
for given datasets which names matche ADAM datasets names.
}
\details{
This function checks if there were keys added to all data sets
}
\examples{

ADSL <- teal.data::example_cdisc_data("ADSL")
ADTTE <- teal.data::example_cdisc_data("ADTTE")

cdisc_data(
  ADSL = ADSL,
  ADTTE = ADTTE,
  code = quote({
    ADSL <- teal.data::example_cdisc_data("ADSL")
    ADTTE <- teal.data::example_cdisc_data("ADTTE")
  }),
  join_keys = join_keys(
    join_key("ADSL", "ADTTE", c("STUDYID" = "STUDYID", "USUBJID" = "USUBJID"))
  )
)
}
