% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdisc_data.R
\name{cdisc_data_file}
\alias{cdisc_data_file}
\title{Load \code{TealData} object from a file}
\usage{
cdisc_data_file(path, code = get_code(path))
}
\arguments{
\item{path}{A (\code{connection}) or a (\code{character})\cr
string giving the pathname of the file or URL to read from. "" indicates the connection \code{stdin}.}

\item{code}{(\code{character})\cr
reproducible code to re-create object}
}
\value{
\code{TealData} object
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
\examples{
file_example <- tempfile(fileext = ".R")
writeLines(
  text = c(
    "# code>
     ADSL <- teal.data::example_cdisc_data('ADSL')
     ADTTE <- teal.data::example_cdisc_data('ADTTE')

     cdisc_data(
          cdisc_dataset(\"ADSL\", ADSL), cdisc_dataset(\"ADTTE\", ADTTE),
          code = \"ADSL <- teal.data::example_cdisc_data('ADSL')
                  ADTTE <- teal.data::example_cdisc_data('ADTTE')\",
          check = FALSE
     )
     # <code"
  ),
  con = file_example
)

cdisc_data_file(file_example)
}
