% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CDISCTealDataConnector.R
\name{CDISCTealDataConnector}
\alias{CDISCTealDataConnector}
\title{Manage multiple and \code{TealDatasetConnector} of the same type.}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
Class manages \code{TealDatasetConnector} to specify additional dynamic arguments and to
open/close connection.
}
\section{Super classes}{
\code{\link[teal.data:TealDataAbstract]{teal.data::TealDataAbstract}} -> \code{\link[teal.data:TealDataConnector]{teal.data::TealDataConnector}} -> \code{CDISCTealDataConnector}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{CDISCTealDataConnector$new()}}
\item \href{#method-get_parent}{\code{CDISCTealDataConnector$get_parent()}}
\item \href{#method-clone}{\code{CDISCTealDataConnector$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="teal.data" data-topic="TealDataAbstract" data-id="check">}\href{../../teal.data/html/TealDataAbstract.html#method-check}{\code{teal.data::TealDataAbstract$check()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="teal.data" data-topic="TealDataAbstract" data-id="check_reproducibility">}\href{../../teal.data/html/TealDataAbstract.html#method-check_reproducibility}{\code{teal.data::TealDataAbstract$check_reproducibility()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="teal.data" data-topic="TealDataAbstract" data-id="execute_mutate">}\href{../../teal.data/html/TealDataAbstract.html#method-execute_mutate}{\code{teal.data::TealDataAbstract$execute_mutate()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="teal.data" data-topic="TealDataAbstract" data-id="get_check">}\href{../../teal.data/html/TealDataAbstract.html#method-get_check}{\code{teal.data::TealDataAbstract$get_check()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="teal.data" data-topic="TealDataAbstract" data-id="get_check_result">}\href{../../teal.data/html/TealDataAbstract.html#method-get_check_result}{\code{teal.data::TealDataAbstract$get_check_result()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="teal.data" data-topic="TealDataAbstract" data-id="get_code">}\href{../../teal.data/html/TealDataAbstract.html#method-get_code}{\code{teal.data::TealDataAbstract$get_code()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="teal.data" data-topic="TealDataAbstract" data-id="get_datanames">}\href{../../teal.data/html/TealDataAbstract.html#method-get_datanames}{\code{teal.data::TealDataAbstract$get_datanames()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="teal.data" data-topic="TealDataAbstract" data-id="get_dataset">}\href{../../teal.data/html/TealDataAbstract.html#method-get_dataset}{\code{teal.data::TealDataAbstract$get_dataset()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="teal.data" data-topic="TealDataAbstract" data-id="get_datasets">}\href{../../teal.data/html/TealDataAbstract.html#method-get_datasets}{\code{teal.data::TealDataAbstract$get_datasets()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="teal.data" data-topic="TealDataAbstract" data-id="get_items">}\href{../../teal.data/html/TealDataAbstract.html#method-get_items}{\code{teal.data::TealDataAbstract$get_items()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="teal.data" data-topic="TealDataAbstract" data-id="is_pulled">}\href{../../teal.data/html/TealDataAbstract.html#method-is_pulled}{\code{teal.data::TealDataAbstract$is_pulled()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="teal.data" data-topic="TealDataAbstract" data-id="mutate_dataset">}\href{../../teal.data/html/TealDataAbstract.html#method-mutate_dataset}{\code{teal.data::TealDataAbstract$mutate_dataset()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="teal.data" data-topic="TealDataAbstract" data-id="reassign_datasets_vars">}\href{../../teal.data/html/TealDataAbstract.html#method-reassign_datasets_vars}{\code{teal.data::TealDataAbstract$reassign_datasets_vars()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="teal.data" data-topic="TealDataAbstract" data-id="set_check">}\href{../../teal.data/html/TealDataAbstract.html#method-set_check}{\code{teal.data::TealDataAbstract$set_check()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="teal.data" data-topic="TealDataAbstract" data-id="set_pull_code">}\href{../../teal.data/html/TealDataAbstract.html#method-set_pull_code}{\code{teal.data::TealDataAbstract$set_pull_code()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="teal.data" data-topic="TealDataConnector" data-id="get_code_class">}\href{../../teal.data/html/TealDataConnector.html#method-get_code_class}{\code{teal.data::TealDataConnector$get_code_class()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="teal.data" data-topic="TealDataConnector" data-id="get_connection">}\href{../../teal.data/html/TealDataConnector.html#method-get_connection}{\code{teal.data::TealDataConnector$get_connection()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="teal.data" data-topic="TealDataConnector" data-id="get_preopen_server">}\href{../../teal.data/html/TealDataConnector.html#method-get_preopen_server}{\code{teal.data::TealDataConnector$get_preopen_server()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="teal.data" data-topic="TealDataConnector" data-id="get_server">}\href{../../teal.data/html/TealDataConnector.html#method-get_server}{\code{teal.data::TealDataConnector$get_server()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="teal.data" data-topic="TealDataConnector" data-id="get_ui">}\href{../../teal.data/html/TealDataConnector.html#method-get_ui}{\code{teal.data::TealDataConnector$get_ui()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="teal.data" data-topic="TealDataConnector" data-id="is_failed">}\href{../../teal.data/html/TealDataConnector.html#method-is_failed}{\code{teal.data::TealDataConnector$is_failed()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="teal.data" data-topic="TealDataConnector" data-id="launch">}\href{../../teal.data/html/TealDataConnector.html#method-launch}{\code{teal.data::TealDataConnector$launch()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="teal.data" data-topic="TealDataConnector" data-id="mutate">}\href{../../teal.data/html/TealDataConnector.html#method-mutate}{\code{teal.data::TealDataConnector$mutate()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="teal.data" data-topic="TealDataConnector" data-id="print">}\href{../../teal.data/html/TealDataConnector.html#method-print}{\code{teal.data::TealDataConnector$print()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="teal.data" data-topic="TealDataConnector" data-id="pull">}\href{../../teal.data/html/TealDataConnector.html#method-pull}{\code{teal.data::TealDataConnector$pull()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="teal.data" data-topic="TealDataConnector" data-id="set_preopen_server">}\href{../../teal.data/html/TealDataConnector.html#method-set_preopen_server}{\code{teal.data::TealDataConnector$set_preopen_server()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="teal.data" data-topic="TealDataConnector" data-id="set_pull_args">}\href{../../teal.data/html/TealDataConnector.html#method-set_pull_args}{\code{teal.data::TealDataConnector$set_pull_args()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="teal.data" data-topic="TealDataConnector" data-id="set_server">}\href{../../teal.data/html/TealDataConnector.html#method-set_server}{\code{teal.data::TealDataConnector$set_server()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="teal.data" data-topic="TealDataConnector" data-id="set_ui">}\href{../../teal.data/html/TealDataConnector.html#method-set_ui}{\code{teal.data::TealDataConnector$set_ui()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{CDISCTealDataConnector} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CDISCTealDataConnector$new(connection, connectors)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{connection}}{(\code{TealDataConnection})\cr
connection to data source}

\item{\code{connectors}}{(\code{list} of \code{TealDatasetConnector} elements)\cr
list with dataset connectors}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_parent"></a>}}
\if{latex}{\out{\hypertarget{method-get_parent}{}}}
\subsection{Method \code{get_parent()}}{
Get all datasets parent names
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CDISCTealDataConnector$get_parent()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
(named \code{list}) with dataset name and its corresponding parent dataset name
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CDISCTealDataConnector$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
