% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CDISCTealDataset.R
\name{CDISCTealDataset}
\alias{CDISCTealDataset}
\title{R6 Class representing a dataset with parent attribute}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
Any \code{data.frame} object can be stored inside this object.

The difference compared to \code{TealDataset} class is a parent field that
indicates name of the parent dataset. Note that the parent field might
be empty (i.e. \code{character(0)}).
}
\examples{
x <- cdisc_dataset(
  dataname = "XYZ",
  x = data.frame(x = c(1, 2), y = c("a", "b"), stringsAsFactors = FALSE),
  keys = "y",
  parent = "ABC",
  code = "XYZ <- data.frame(x = c(1, 2), y = c('aa', 'bb'),
                            stringsAsFactors = FALSE)",
  metadata = list(type = "example")
)

x$ncol
x$get_code()
x$get_dataname()
x$get_keys()
x$get_parent()
}
\section{Super class}{
\code{\link[teal.data:TealDataset]{teal.data::TealDataset}} -> \code{CDISCTealDataset}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CDISCTealDataset-new}{\code{CDISCTealDataset$new()}}
\item \href{#method-CDISCTealDataset-recreate}{\code{CDISCTealDataset$recreate()}}
\item \href{#method-CDISCTealDataset-get_attrs}{\code{CDISCTealDataset$get_attrs()}}
\item \href{#method-CDISCTealDataset-get_parent}{\code{CDISCTealDataset$get_parent()}}
\item \href{#method-CDISCTealDataset-set_parent}{\code{CDISCTealDataset$set_parent()}}
\item \href{#method-CDISCTealDataset-clone}{\code{CDISCTealDataset$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="teal.data" data-topic="TealDataset" data-id="check"><a href='../../teal.data/html/TealDataset.html#method-TealDataset-check'><code>teal.data::TealDataset$check()</code></a></span></li>
<li><span class="pkg-link" data-pkg="teal.data" data-topic="TealDataset" data-id="check_keys"><a href='../../teal.data/html/TealDataset.html#method-TealDataset-check_keys'><code>teal.data::TealDataset$check_keys()</code></a></span></li>
<li><span class="pkg-link" data-pkg="teal.data" data-topic="TealDataset" data-id="get_character_colnames"><a href='../../teal.data/html/TealDataset.html#method-TealDataset-get_character_colnames'><code>teal.data::TealDataset$get_character_colnames()</code></a></span></li>
<li><span class="pkg-link" data-pkg="teal.data" data-topic="TealDataset" data-id="get_code"><a href='../../teal.data/html/TealDataset.html#method-TealDataset-get_code'><code>teal.data::TealDataset$get_code()</code></a></span></li>
<li><span class="pkg-link" data-pkg="teal.data" data-topic="TealDataset" data-id="get_code_class"><a href='../../teal.data/html/TealDataset.html#method-TealDataset-get_code_class'><code>teal.data::TealDataset$get_code_class()</code></a></span></li>
<li><span class="pkg-link" data-pkg="teal.data" data-topic="TealDataset" data-id="get_colnames"><a href='../../teal.data/html/TealDataset.html#method-TealDataset-get_colnames'><code>teal.data::TealDataset$get_colnames()</code></a></span></li>
<li><span class="pkg-link" data-pkg="teal.data" data-topic="TealDataset" data-id="get_column_labels"><a href='../../teal.data/html/TealDataset.html#method-TealDataset-get_column_labels'><code>teal.data::TealDataset$get_column_labels()</code></a></span></li>
<li><span class="pkg-link" data-pkg="teal.data" data-topic="TealDataset" data-id="get_dataname"><a href='../../teal.data/html/TealDataset.html#method-TealDataset-get_dataname'><code>teal.data::TealDataset$get_dataname()</code></a></span></li>
<li><span class="pkg-link" data-pkg="teal.data" data-topic="TealDataset" data-id="get_datanames"><a href='../../teal.data/html/TealDataset.html#method-TealDataset-get_datanames'><code>teal.data::TealDataset$get_datanames()</code></a></span></li>
<li><span class="pkg-link" data-pkg="teal.data" data-topic="TealDataset" data-id="get_dataset"><a href='../../teal.data/html/TealDataset.html#method-TealDataset-get_dataset'><code>teal.data::TealDataset$get_dataset()</code></a></span></li>
<li><span class="pkg-link" data-pkg="teal.data" data-topic="TealDataset" data-id="get_dataset_label"><a href='../../teal.data/html/TealDataset.html#method-TealDataset-get_dataset_label'><code>teal.data::TealDataset$get_dataset_label()</code></a></span></li>
<li><span class="pkg-link" data-pkg="teal.data" data-topic="TealDataset" data-id="get_factor_colnames"><a href='../../teal.data/html/TealDataset.html#method-TealDataset-get_factor_colnames'><code>teal.data::TealDataset$get_factor_colnames()</code></a></span></li>
<li><span class="pkg-link" data-pkg="teal.data" data-topic="TealDataset" data-id="get_hash"><a href='../../teal.data/html/TealDataset.html#method-TealDataset-get_hash'><code>teal.data::TealDataset$get_hash()</code></a></span></li>
<li><span class="pkg-link" data-pkg="teal.data" data-topic="TealDataset" data-id="get_join_keys"><a href='../../teal.data/html/TealDataset.html#method-TealDataset-get_join_keys'><code>teal.data::TealDataset$get_join_keys()</code></a></span></li>
<li><span class="pkg-link" data-pkg="teal.data" data-topic="TealDataset" data-id="get_keys"><a href='../../teal.data/html/TealDataset.html#method-TealDataset-get_keys'><code>teal.data::TealDataset$get_keys()</code></a></span></li>
<li><span class="pkg-link" data-pkg="teal.data" data-topic="TealDataset" data-id="get_metadata"><a href='../../teal.data/html/TealDataset.html#method-TealDataset-get_metadata'><code>teal.data::TealDataset$get_metadata()</code></a></span></li>
<li><span class="pkg-link" data-pkg="teal.data" data-topic="TealDataset" data-id="get_mutate_code_class"><a href='../../teal.data/html/TealDataset.html#method-TealDataset-get_mutate_code_class'><code>teal.data::TealDataset$get_mutate_code_class()</code></a></span></li>
<li><span class="pkg-link" data-pkg="teal.data" data-topic="TealDataset" data-id="get_mutate_vars"><a href='../../teal.data/html/TealDataset.html#method-TealDataset-get_mutate_vars'><code>teal.data::TealDataset$get_mutate_vars()</code></a></span></li>
<li><span class="pkg-link" data-pkg="teal.data" data-topic="TealDataset" data-id="get_ncol"><a href='../../teal.data/html/TealDataset.html#method-TealDataset-get_ncol'><code>teal.data::TealDataset$get_ncol()</code></a></span></li>
<li><span class="pkg-link" data-pkg="teal.data" data-topic="TealDataset" data-id="get_nrow"><a href='../../teal.data/html/TealDataset.html#method-TealDataset-get_nrow'><code>teal.data::TealDataset$get_nrow()</code></a></span></li>
<li><span class="pkg-link" data-pkg="teal.data" data-topic="TealDataset" data-id="get_numeric_colnames"><a href='../../teal.data/html/TealDataset.html#method-TealDataset-get_numeric_colnames'><code>teal.data::TealDataset$get_numeric_colnames()</code></a></span></li>
<li><span class="pkg-link" data-pkg="teal.data" data-topic="TealDataset" data-id="get_raw_data"><a href='../../teal.data/html/TealDataset.html#method-TealDataset-get_raw_data'><code>teal.data::TealDataset$get_raw_data()</code></a></span></li>
<li><span class="pkg-link" data-pkg="teal.data" data-topic="TealDataset" data-id="get_row_labels"><a href='../../teal.data/html/TealDataset.html#method-TealDataset-get_row_labels'><code>teal.data::TealDataset$get_row_labels()</code></a></span></li>
<li><span class="pkg-link" data-pkg="teal.data" data-topic="TealDataset" data-id="get_rownames"><a href='../../teal.data/html/TealDataset.html#method-TealDataset-get_rownames'><code>teal.data::TealDataset$get_rownames()</code></a></span></li>
<li><span class="pkg-link" data-pkg="teal.data" data-topic="TealDataset" data-id="get_var_r6"><a href='../../teal.data/html/TealDataset.html#method-TealDataset-get_var_r6'><code>teal.data::TealDataset$get_var_r6()</code></a></span></li>
<li><span class="pkg-link" data-pkg="teal.data" data-topic="TealDataset" data-id="get_vars"><a href='../../teal.data/html/TealDataset.html#method-TealDataset-get_vars'><code>teal.data::TealDataset$get_vars()</code></a></span></li>
<li><span class="pkg-link" data-pkg="teal.data" data-topic="TealDataset" data-id="is_mutate_delayed"><a href='../../teal.data/html/TealDataset.html#method-TealDataset-is_mutate_delayed'><code>teal.data::TealDataset$is_mutate_delayed()</code></a></span></li>
<li><span class="pkg-link" data-pkg="teal.data" data-topic="TealDataset" data-id="is_pulled"><a href='../../teal.data/html/TealDataset.html#method-TealDataset-is_pulled'><code>teal.data::TealDataset$is_pulled()</code></a></span></li>
<li><span class="pkg-link" data-pkg="teal.data" data-topic="TealDataset" data-id="merge_join_keys"><a href='../../teal.data/html/TealDataset.html#method-TealDataset-merge_join_keys'><code>teal.data::TealDataset$merge_join_keys()</code></a></span></li>
<li><span class="pkg-link" data-pkg="teal.data" data-topic="TealDataset" data-id="mutate"><a href='../../teal.data/html/TealDataset.html#method-TealDataset-mutate'><code>teal.data::TealDataset$mutate()</code></a></span></li>
<li><span class="pkg-link" data-pkg="teal.data" data-topic="TealDataset" data-id="mutate_join_keys"><a href='../../teal.data/html/TealDataset.html#method-TealDataset-mutate_join_keys'><code>teal.data::TealDataset$mutate_join_keys()</code></a></span></li>
<li><span class="pkg-link" data-pkg="teal.data" data-topic="TealDataset" data-id="print"><a href='../../teal.data/html/TealDataset.html#method-TealDataset-print'><code>teal.data::TealDataset$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="teal.data" data-topic="TealDataset" data-id="reassign_datasets_vars"><a href='../../teal.data/html/TealDataset.html#method-TealDataset-reassign_datasets_vars'><code>teal.data::TealDataset$reassign_datasets_vars()</code></a></span></li>
<li><span class="pkg-link" data-pkg="teal.data" data-topic="TealDataset" data-id="set_code"><a href='../../teal.data/html/TealDataset.html#method-TealDataset-set_code'><code>teal.data::TealDataset$set_code()</code></a></span></li>
<li><span class="pkg-link" data-pkg="teal.data" data-topic="TealDataset" data-id="set_dataset_label"><a href='../../teal.data/html/TealDataset.html#method-TealDataset-set_dataset_label'><code>teal.data::TealDataset$set_dataset_label()</code></a></span></li>
<li><span class="pkg-link" data-pkg="teal.data" data-topic="TealDataset" data-id="set_join_keys"><a href='../../teal.data/html/TealDataset.html#method-TealDataset-set_join_keys'><code>teal.data::TealDataset$set_join_keys()</code></a></span></li>
<li><span class="pkg-link" data-pkg="teal.data" data-topic="TealDataset" data-id="set_keys"><a href='../../teal.data/html/TealDataset.html#method-TealDataset-set_keys'><code>teal.data::TealDataset$set_keys()</code></a></span></li>
<li><span class="pkg-link" data-pkg="teal.data" data-topic="TealDataset" data-id="set_vars"><a href='../../teal.data/html/TealDataset.html#method-TealDataset-set_vars'><code>teal.data::TealDataset$set_vars()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CDISCTealDataset-new"></a>}}
\if{latex}{\out{\hypertarget{method-CDISCTealDataset-new}{}}}
\subsection{Method \code{new()}}{
Create a new object of \code{CDISCTealDataset} class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CDISCTealDataset$new(
  dataname,
  x,
  keys,
  parent,
  code = character(0),
  label = character(0),
  vars = list(),
  metadata = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dataname}}{(\code{character})\cr
A given name for the dataset it may not contain spaces}

\item{\code{x}}{(\code{data.frame})\cr}

\item{\code{keys}}{(\code{character})\cr
vector with primary keys}

\item{\code{parent}}{optional, (\code{character}) \cr
parent dataset name}

\item{\code{code}}{(\code{character})\cr
A character string defining the code needed to produce the data set in \code{x}}

\item{\code{label}}{(\code{character})\cr
Label to describe the dataset}

\item{\code{vars}}{(named \code{list})) \cr
In case when this object code depends on other \code{TealDataset} object(s) or
other constant value, this/these object(s) should be included as named
element(s) of the list. For example if this object code needs \code{ADSL}
object we should specify \verb{vars = list(ADSL = <adsl object>)}.
It is recommended to include \code{TealDataset} or \code{TealDatasetConnector} objects to
the \code{vars} list to preserve reproducibility. Please note that \code{vars}
are included to this object as local \code{vars} and they cannot be modified
within another dataset.}

\item{\code{metadata}}{(named \code{list} or \code{NULL}) \cr
Field containing metadata about the dataset. Each element of the list
should be atomic and length one.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CDISCTealDataset-recreate"></a>}}
\if{latex}{\out{\hypertarget{method-CDISCTealDataset-recreate}{}}}
\subsection{Method \code{recreate()}}{
Recreate a dataset with its current attributes
This is useful way to have access to class initialize method basing on class object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CDISCTealDataset$recreate(
  dataname = self$get_dataname(),
  x = self$get_raw_data(),
  keys = self$get_keys(),
  parent = self$get_parent(),
  code = private$code,
  label = self$get_dataset_label(),
  vars = list(),
  metadata = self$get_metadata()
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dataname}}{(\code{character})\cr
A given name for the dataset it may not contain spaces}

\item{\code{x}}{(\code{data.frame})\cr}

\item{\code{keys}}{(\code{character})\cr
vector with primary keys}

\item{\code{parent}}{optional, (\code{character}) \cr
parent dataset name}

\item{\code{code}}{(\code{character})\cr
A character string defining the code needed to produce the data set in \code{x}}

\item{\code{label}}{(\code{character})\cr
Label to describe the dataset}

\item{\code{vars}}{(named \code{list})) \cr
In case when this object code depends on other \code{TealDataset} object(s) or
other constant value, this/these object(s) should be included as named
element(s) of the list. For example if this object code needs \code{ADSL}
object we should specify \verb{vars = list(ADSL = <adsl object>)}.
It is recommended to include \code{TealDataset} or \code{TealDatasetConnector} objects to
the \code{vars} list to preserve reproducibility. Please note that \code{vars}
are included to this object as local \code{vars} and they cannot be modified
within another dataset.}

\item{\code{metadata}}{(named \code{list} or \code{NULL}) \cr
Field containing metadata about the dataset. Each element of the list
should be atomic and length one.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a new object of \code{CDISCTealDataset} class
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CDISCTealDataset-get_attrs"></a>}}
\if{latex}{\out{\hypertarget{method-CDISCTealDataset-get_attrs}{}}}
\subsection{Method \code{get_attrs()}}{
Get all dataset attributes
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CDISCTealDataset$get_attrs()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
(named \code{list}) with dataset attributes
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CDISCTealDataset-get_parent"></a>}}
\if{latex}{\out{\hypertarget{method-CDISCTealDataset-get_parent}{}}}
\subsection{Method \code{get_parent()}}{
Get parent dataset name
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CDISCTealDataset$get_parent()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
(\code{character}) indicating parent dataname
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CDISCTealDataset-set_parent"></a>}}
\if{latex}{\out{\hypertarget{method-CDISCTealDataset-set_parent}{}}}
\subsection{Method \code{set_parent()}}{
Set parent dataset name
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CDISCTealDataset$set_parent(parent)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{parent}}{(\code{character}) indicating parent dataname}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
(\code{self}) invisibly for chaining
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CDISCTealDataset-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CDISCTealDataset-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CDISCTealDataset$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
