% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parents.R
\name{check_parent_child}
\alias{check_parent_child}
\title{Check if parent/child are valid}
\usage{
check_parent_child(join_keys_obj)
}
\description{
Check if parent/child are valid
}
\examples{
jk <- new_join_keys()
jk["ds1", "ds2"] <- character(0)
parents(jk) <- list(ds1 = "ds2")
check_parent_child(jk)
}
\keyword{internal}
