% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatters_var_labels.R
\name{col_relabel}
\alias{col_relabel}
\title{Copy and Change Variable Labels of a \code{data.frame}}
\source{
This function was taken 1-1 from
\href{https://cran.r-project.org/web/packages/formatters/index.html}{formatters} package, to reduce the complexity of
the dependency tree.
}
\usage{
col_relabel(x, ...)
}
\arguments{
\item{x}{a \code{data.frame} object}

\item{...}{name-value pairs, where name corresponds to a variable name in
\code{x} and the value to the new variable label}
}
\value{
a copy of \code{x} with changed labels according to \code{...}
}
\description{
Relabel a subset of the variables
}
\examples{
x <- col_relabel(iris, Sepal.Length = "Sepal Length of iris flower")
col_labels(x)

}
\seealso{
\code{\link[=col_labels]{col_labels()}} \code{\link{col_labels<-}}
}
