% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_key.R
\name{dataset_1.JoinKeySet}
\alias{dataset_1.JoinKeySet}
\alias{dataset_2.JoinKeySet}
\alias{keys.JoinKeySet}
\title{Getter for attributes in \code{JoinKeySet} object}
\usage{
dataset_1.JoinKeySet(join_key_object)

dataset_2.JoinKeySet(join_key_object)

keys.JoinKeySet(join_key_object)
}
\arguments{
\item{join_key_object}{(\code{JoinKeySet}) object to retrieve attribute from.}
}
\value{
\code{dataset_1}, \code{dataset_2} or \code{key} as \code{character(1)}
}
\description{
Internal methods for \code{JoinKeySet} operations
}
\keyword{internal}
