% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TealDatasetConnector_constructors.R
\name{dataset_connector_file}
\alias{dataset_connector_file}
\alias{cdisc_dataset_connector_file}
\title{Load \code{TealDatasetConnector} object from a file}
\usage{
dataset_connector_file(path)

cdisc_dataset_connector_file(path)
}
\arguments{
\item{path}{(\code{character}) string giving the pathname of the file to read from.}
}
\value{
\code{TealDatasetConnector} object

\code{CDISCTealDatasetConnector} object
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
\details{
Please note that the script has to end with a call creating desired object. The error will
be raised otherwise.

Please note that the script has to end with a call creating desired object. The error will
be raised otherwise.
}
\examples{
# simple example
library(magrittr)
file_example <- tempfile(fileext = ".R")
writeLines(
  text = c(
    "library(teal.data)

     pull_callable <- callable_function(teal.data::example_cdisc_data) \%>\%
       set_args(list(dataname = \"ADSL\"))
     dataset_connector(\"ADSL\", pull_callable, get_cdisc_keys(\"ADSL\"))"
  ),
  con = file_example
)
x <- dataset_connector_file(file_example)
get_code(x)
# simple example
file_example <- tempfile(fileext = ".R")
writeLines(
  text = c(
    "library(teal.data)

     pull_callable <- callable_function(teal.data::example_cdisc_data) \%>\%
       set_args(list(dataname = \"ADSL\"))
     cdisc_dataset_connector(\"ADSL\", pull_callable, get_cdisc_keys(\"ADSL\"))"
  ),
  con = file_example
)
x <- cdisc_dataset_connector_file(file_example)
get_code(x)
}
