% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddl.R
\name{ddl_eval_substitute}
\alias{ddl_eval_substitute}
\title{Substitute and evaluate ddl code}
\usage{
ddl_eval_substitute(code, args)
}
\arguments{
\item{code}{(\code{character})\cr
Code to be evaluated and returned to the \code{postprocess_fun}}

\item{args}{(\code{list} named)\cr
Containing elements named after arguments in the code
enclosed in currly brackets ex. \code{{ arg_name }}}
}
\value{
\code{list} of objects being a result of the code evaluation
}
\description{
Substitute and evaluate ddl code
}
\examples{
ddl_eval_substitute("x <- { arg }", list(arg = 1))
ddl_eval_substitute("x <- { arg }", list(arg = "a"))
ddl_eval_substitute("a <- 1; x <- { arg } + 1", list(arg = quote(a)))
ddl_eval_substitute("a <- b", list(b = 1))
}
