% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddl.R
\name{ddl_run}
\alias{ddl_run}
\title{Creates \code{tdata} object}
\usage{
ddl_run(x, online_args = list())
}
\arguments{
\item{online_args}{(\code{list} named)\cr
Arguments to be substituted in the \code{code} and evaluated. Result of the evaluation
is based on the provided (dynamic) arguments.}
}
\value{
\code{tdata} containing objects created:
\itemize{
\item \code{env} created by the \code{code} substitution and evaluation using
\code{online_args}, while the \code{code}.
\item \code{code} with substituted `offline_args.
\item \code{join_keys} specified in the \code{ddl} object.
}
}
\description{
Resolves arguments and executes custom DDL \code{code}.
Custom \code{code} is substituted by \code{online_args} and evaluated. Then obtained code is
substituted again by \code{offline_args} and passed to the \code{postprocess_fun}.
}
