% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_keys-c.R
\name{.merge_parents}
\alias{.merge_parents}
\title{Merge parents for 2 \code{join_keys} object}
\usage{
.merge_parents(x, y)
}
\arguments{
\item{x, y}{(\code{join_keys}) objects to merge their parents}
}
\value{
a list with parents merged from 2 \code{join_keys}. Not the object itself.
}
\description{
Merge parents for 2 \code{join_keys} object
}
\keyword{internal}
