% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format_expression}
\alias{format_expression}
\title{Format expression to string
Convert any expression to a single character vector}
\usage{
format_expression(code)
}
\arguments{
\item{code}{(\code{language}, \code{expression}, \code{character})}
}
\value{
\code{character(1)}
}
\description{
Format expression to string
Convert any expression to a single character vector
}
\keyword{internal}
