% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatters_var_labels.R
\name{formatters_var_labels<-}
\alias{formatters_var_labels<-}
\title{Set Label Attributes of All Variables in a \code{data.frame}}
\source{
This function was taken 1-1 from
\href{https://cran.r-project.org/web/packages/formatters/index.html}{formatters} package, to reduce the complexity of
the dependency tree.
}
\usage{
formatters_var_labels(x) <- value
}
\arguments{
\item{x}{a \code{data.frame} object}

\item{value}{new variable labels, \code{NA} removes the variable label}
}
\value{
modifies the variable labels of \code{x}
}
\description{
Variable labels can be stored as a \code{label} attribute for each variable.
This functions sets all non-missing (non-NA) variable labels in a \code{data.frame}
}
\examples{
x <- iris
formatters_var_labels(x)
formatters_var_labels(x) <- paste("label for", names(iris))
formatters_var_labels(x)

if (interactive()) {
  View(x) # in RStudio data viewer labels are displayed
}
}
\seealso{
\code{\link[=formatters_var_labels]{formatters_var_labels()}} \code{\link[=formatters_var_relabel]{formatters_var_relabel()}} \code{\link[=formatters_with_label]{formatters_with_label()}}
}
