% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatters_var_labels.R
\name{formatters_with_label}
\alias{formatters_with_label}
\title{Return an object with a label attribute}
\source{
This function was taken 1-1 from
\href{https://cran.r-project.org/web/packages/formatters/index.html}{formatters} package, to reduce the complexity of
the dependency tree.
}
\usage{
formatters_with_label(x, label)
}
\arguments{
\item{x}{an object}

\item{label}{label attribute to to attached to  \code{x}}
}
\value{
\code{x} labeled by \code{label}. Note: the exact mechanism of labeling should be
considered an internal implementation detail, but the label will always be retrieved via \code{obj_label}.
}
\description{
Return an object with a label attribute
}
\examples{
x <- formatters_with_label(c(1, 2, 3), label = "Test")
obj_label(x)
}
\seealso{
\code{\link[=formatters_var_labels]{formatters_var_labels()}} \code{\link[=formatters_var_relabel]{formatters_var_relabel()}} \code{\link{formatters_var_labels<-}}
}
