% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_attrs.R
\name{get_attrs}
\alias{get_attrs}
\alias{get_attrs.TealDataset}
\title{Get dataset attributes}
\usage{
get_attrs(x)

\method{get_attrs}{TealDataset}(x)
}
\arguments{
\item{x}{an object of (\code{TealDataset}) class}
}
\value{
named \code{list} of object attributes
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
Get dataset attributes in form of named list.
}
\examples{
# TealDataset --------

ADSL <- example_cdisc_data("ADSL")

x1 <- dataset("ADSL", x = ADSL, label = "custom label")
get_attrs(x1)

x2 <- dataset(
  "ADSL",
  x = ADSL,
  keys = get_cdisc_keys("ADSL"),
  label = "custom label"
)
get_attrs(x2)

# CDISCTealDataset --------

ADSL <- example_cdisc_data("ADSL")
x3 <- cdisc_dataset(
  "ADSL",
  x = ADSL,
  keys = get_cdisc_keys("ADSL"),
  label = "custom label"
)
get_attrs(x3)
}
