% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-get_code_dependency.R
\name{get_code_dependency}
\alias{get_code_dependency}
\title{Return the lines of code (with side-effects) needed to reproduce the object}
\usage{
get_code_dependency(code, names)
}
\arguments{
\item{code}{\code{character} with the code.}

\item{names}{\code{character} vector of object names.}
}
\value{
\code{character} vector of elements of \code{code} calls that were required to build the side-effects and
influencing objects having and impact on the \code{object}
}
\description{
Return the lines of code (with side-effects) needed to reproduce the object
}
\section{Notes}{

This function assumes that object relationships are established using the \verb{<-}, \code{=}, or \verb{->} assignment
operators. It does not support other object creation methods like \code{assign} or \verb{<<-}, nor non-standard-evaluation
methods. To specify relationships between side-effects and objects, you can use the comment tag
\verb{# @linksto object_name} at the end of a line where the side-effect occurs.
}

\keyword{internal}
