% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dataset.R
\name{get_dataset}
\alias{get_dataset}
\alias{get_dataset.TealDatasetConnector}
\alias{get_dataset.TealDataset}
\alias{get_dataset.TealDataAbstract}
\title{Get dataset from \code{TealDatasetConnector}}
\usage{
get_dataset(x, dataname)

\method{get_dataset}{TealDatasetConnector}(x, dataname = NULL)

\method{get_dataset}{TealDataset}(x, dataname = NULL)

\method{get_dataset}{TealDataAbstract}(x, dataname = NULL)
}
\arguments{
\item{x}{(\code{TealDatasetConnector} or \code{TealDatasetConnector} or \code{TealDataAbstract})}

\item{dataname}{(\code{character}) a name of dataset to be retrieved}
}
\value{
(\code{TealDataset})
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Get dataset from \code{TealDatasetConnector}
}
\details{
See \code{help(TealDataConnector)} and \code{help(TealData)} for more complex examples.
}
\examples{

# TealDatasetConnector --------
pull_fun_adae <- callable_function(teal.data::example_cdisc_data) \%>\%
  set_args(list(dataname = "ADAE"))

ADSL <- teal.data::example_cdisc_data("ADSL")

dc <- dataset_connector(
  dataname = "ADAE", pull_callable = pull_fun_adae,
  keys = get_cdisc_keys("ADSL")
)

\dontrun{
load_dataset(dc)
get_dataset(dc)
}


# TealDataset --------
ADSL <- example_cdisc_data("ADSL")
x <- dataset("ADSL", ADSL)

get_dataset(x)

# TealData  (not containing connectors) --------
adsl <- cdisc_dataset(
  dataname = "ADSL",
  x = example_cdisc_data("ADSL"),
  code = "library(teal.data)\nADSL <- example_cdisc_data(\"ADSL\")"
)

adae <- cdisc_dataset(
  dataname = "ADAE",
  x = example_cdisc_data("ADAE"),
  code = "library(teal.data)\nADAE <- example_cdisc_data(\"ADAE\")"
)

rd <- teal.data:::TealData$new(adsl, adae)
get_dataset(rd, "ADSL")
}
