% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_keys.R
\name{get_join_key}
\alias{get_join_key}
\title{Get value of a single relationship pair}
\usage{
get_join_key(join_keys_obj, dataset_1 = NULL, dataset_2 = NULL)
}
\arguments{
\item{join_keys_obj}{(\code{JoinKeys}) object that holds the relationship keys.}

\item{dataset_1}{(\code{character(1)}) one of the datasets to retrieve keys (
order of the datasets is irrelevant).}

\item{dataset_2}{(\code{character(1)}) the other dataset to retrieve keys (the
order of the datasets is irrelevant).}
}
\value{
Character vector with keys or (if one of the datasets is omitted) a
list of relationship pairs. If both datasets are omitted it returens the
\code{JoinKeys} object
}
\description{
Get value of a single relationship pair
}
\keyword{internal}
