% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_key_duplicates.R
\name{get_key_duplicates_util}
\alias{get_key_duplicates_util}
\title{Creates a duplicate keys information summary.}
\usage{
get_key_duplicates_util(dataframe, keys)
}
\arguments{
\item{dataframe}{dataframe}

\item{keys}{\code{character} vector of variable names consisting the key to the \code{data.frame}}
}
\value{
\code{data.frame} with a duplicate keys information summary
}
\description{
Creates a duplicate keys information summary.
}
\details{
Accepts a list of variable names - \code{keys}, which are treated as the
key to the \code{data.frame} argument. An instance of duplicated key is
defined as two rows, which have the same values in columns defined by \code{keys}.
Per each key value with duplicates returns a row in a \code{tibble}. The return table
has columns corresponding to the variable names passed in \code{keys} and
two additional columns: \code{rows} and \code{n}, which provide
information about row numbers of the original dataframe, which contain duplicated keys
and total duplicates counts.
}
\examples{
df <- data.frame(
  a = c("a", "a", "b", "b", "c"),
  b = c(1, 2, 3, 3, 4),
  c = c(1, 2, 3, 4, 5)
)
res <- teal.data:::get_key_duplicates_util(df, keys = c("a", "b"))
print(res) # duplicated keys are in rows 3 and 4
}
\seealso{
\link{get_key_duplicates}
}
\keyword{internal}
