% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_keys.R, R/new_join_keys.R
\name{get_keys}
\alias{get_keys}
\alias{get_keys.TealDataset}
\alias{get_keys.TealDatasetConnector}
\alias{get_keys.TealDataAbstract}
\alias{join_keys<-}
\alias{join_keys<-.Placeholder}
\title{Get dataset primary keys}
\usage{
get_keys(x, ...)

\method{get_keys}{TealDataset}(x, ...)

\method{get_keys}{TealDatasetConnector}(x, ...)

\method{get_keys}{TealDataAbstract}(x, dataname, ...)

join_keys(join_keys_obj) <- value

\method{join_keys}{Placeholder}(join_keys_obj) <- value
}
\arguments{
\item{x}{an object of \code{TealDataset} or \code{TealDatasetConnector} class}

\item{...}{not used, only for support of S3}

\item{dataname}{(\code{character}) name of dataset to return keys for}

\item{value}{(\code{JoinKeySet} or list of \code{JoinKeySet}) relationship pairs to add
to \code{JoinKeys} list.}

\item{data}{(\code{JoinKeys}) empty object to set the new relationship pairs.}
}
\value{
(\code{character}) vector of column names
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
Get dataset primary keys
}
\examples{
# TealDataset --------

get_keys(
  dataset(
    "ADSL",
    teal.data::example_cdisc_data("ADSL"),
    keys = get_cdisc_keys("ADSL"),
    code = "ADSL <- teal.data::example_cdisc_data(\"ADSL\")"
  )
)
# TealDatasetConnector --------
library(magrittr)
pull_fun_adsl <- callable_function(teal.data::example_cdisc_data) \%>\%
  set_args(list(dataname = "ADAE"))
get_keys(
  dataset_connector(
    "ADSL",
    pull_fun_adsl,
    keys = get_cdisc_keys("ADSL"),
  )
)
# TealData --------

get_keys(
  teal_data(
    dataset("x", data.frame(x1 = 1:10, y1 = 11:20), keys = "x1"),
    dataset("y", data.frame(x2 = 1:10, y2 = 11:20), keys = "x2")
  ),
  "x"
)
jk <- new_join_keys()
join_keys(jk)
join_keys(jk) <- join_key("ds1", "ds2", "some_col2")
}
