% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{get_labels}
\alias{get_labels}
\title{Deprecated function to get label attributes of variables in a \code{data.frame}}
\usage{
get_labels(...)
}
\arguments{
\item{...}{any arguments.}
}
\value{
nothing
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}\cr
This function has been deprecated. Please see \code{\link[=col_labels]{col_labels()}} instead.
}
