% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teal_data.R
\name{get_primary_keys}
\alias{get_primary_keys}
\title{Add primary keys as join_keys to a dataset self}
\usage{
get_primary_keys(data_objects, join_keys)
}
\arguments{
\item{data_objects}{(\code{list}) of TealDataset, TealDatasetConnector or TealDataConnector objects}

\item{join_keys}{(\code{JoinKeys}) object}
}
\description{
Add primary keys as join_keys to a dataset self
}
