% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_data.R
\name{get_raw_data}
\alias{get_raw_data}
\alias{get_raw_data.TealDataset}
\alias{get_raw_data.TealDatasetConnector}
\alias{get_raw_data.TealDataAbstract}
\title{Retrieve raw data}
\usage{
get_raw_data(x, dataname = NULL)

\method{get_raw_data}{TealDataset}(x, dataname = NULL)

\method{get_raw_data}{TealDatasetConnector}(x, dataname = NULL)

\method{get_raw_data}{TealDataAbstract}(x, dataname = NULL)
}
\arguments{
\item{x}{(\code{TealDataset}, \code{TealDatasetConnector}, \code{TealDataAbstract})\cr
object}

\item{dataname}{(\code{character})\cr
Name of dataset to return raw data for.}
}
\value{
\code{data.frame} with the raw data inserted into the R6 objects. In case of
\code{TealDataAbstract}, \code{list} of \code{data.frame} can be returned
if user doesn't specify \code{dataname} - (\code{get_raw_data} from all datasets).
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
\examples{

# TealDataset ---------
library(scda)
ADSL <- synthetic_cdisc_data("latest")$adsl

x <- dataset(dataname = "ADSL", x = ADSL)
get_raw_data(x)

# TealDatasetConnector ---------
library(scda)
pull_fun_adsl <- callable_function(
  function() {
    synthetic_cdisc_data("latest")$adsl
  }
)
dc <- dataset_connector("ADSL", pull_fun_adsl)
load_dataset(dc)
get_raw_data(dc)

# TealData ----------------
library(scda)
latest_data <- synthetic_cdisc_data("latest")
adsl <- cdisc_dataset(
  dataname = "ADSL",
  x = latest_data$adsl,
  code = "library(scda)\nADSL <- synthetic_cdisc_data(\"latest\")$adsl"
)

adtte <- cdisc_dataset(
  dataname = "ADTTE",
  x = latest_data$adtte,
  code = "library(scda)\nADTTE <- synthetic_cdisc_data(\"latest\")$adtte"
)

rd <- teal.data:::TealData$new(adsl, adtte)
get_raw_data(rd)

# TealDataConnector --------
library(scda)
adsl <- scda_cdisc_dataset_connector(dataname = "ADSL", "adsl")
adlb <- scda_cdisc_dataset_connector(dataname = "ADLB", "adlb")

open_fun <- callable_function(library)
open_fun$set_args(list(package = "scda"))

con <- data_connection(open_fun = open_fun)
con$set_open_server(
  function(id, connection) {
    moduleServer(
      id = id,
      module = function(input, output, session) {
        connection$open(try = TRUE)
        return(invisible(connection))
      }
    )
  }
)

rdc <- relational_data_connector(connection = con, connectors = list(adsl, adlb))

rdc$set_ui(
  function(id, connection, connectors) {
    ns <- NS(id)
    tagList(
      connection$get_open_ui(ns("open_connection")),
      textInput(ns("name"), p("Choose", code("scda data version")), value = "latest"),
      do.call(
        what = "tagList",
        args = lapply(
          connectors,
          function(connector) {
            div(
              connector$get_ui(
                id = ns(connector$get_dataname())
              ),
              br()
            )
          }
        )
      )
    )
  }
)

rdc$set_server(
  function(id, connection, connectors) {
    moduleServer(
      id = id,
      module = function(input, output, session) {
        # opens connection
        connection$get_open_server()(id = "open_connection", connection = connection)
        if (connection$is_opened()) {
          for (connector in connectors) {
            set_args(connector, args = list(name = input$name))
            # pull each dataset
            connector$get_server()(id = connector$get_dataname())
            if (connector$is_failed()) {
              break
            }
          }
        }
      }
    )
  }
)

\dontrun{
load_datasets(rdc)
get_raw_data(rdc)
}

# TealData (with connectors) --------
drc <- cdisc_data(rdc)
\dontrun{
get_raw_data(drc)
}
}
