% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddl.R
\name{glue_code}
\alias{glue_code}
\title{Substitute ddl code args}
\usage{
glue_code(code, args)
}
\value{
\code{character}
}
\description{
Substitutes code arguments with \code{args}. Parts of the code
wrapped in curly brackets ex. \code{{ arg_name }} are replaced
with corresponding list elements
}
\examples{
glue_code("x <- { arg }", list(arg = 1))
glue_code("x <- { arg }", list(arg = "a"))
glue_code("a <- 1; x <- { arg } + 1", list(arg = quote(a)))
glue_code(
  "a <- connect(login = { login }, password = { pass})",
  list(
    login = quote(askpass::askpass()),
    password = quote(askpass::askpass())
  )
)
}
