% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_keys.R
\name{join_pair}
\alias{join_pair}
\title{Helper function to add a new pair to a \code{JoinKeys} object}
\usage{
join_pair(join_keys_obj, join_key_obj)
}
\arguments{
\item{join_keys_obj}{(\code{JoinKeys}) Object with existing pairs.}

\item{join_key_obj}{(\code{join_key_set}) relationship pair to add.}
}
\description{
Helper function to add a new pair to a \code{JoinKeys} object
}
\keyword{internal}
