% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_keys-utils.R
\name{length.join_keys}
\alias{length.join_keys}
\title{Length of \code{join_keys} object.}
\usage{
\method{length}{join_keys}(x)
}
\arguments{
\item{x}{an \R object.  For replacement, a vector or factor.}
}
\description{
Length of \code{join_keys} object.
}
