% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JoinKeys.R
\name{length.JoinKeys}
\alias{length.JoinKeys}
\title{Length of an Object}
\usage{
\method{length}{JoinKeys}(x)
}
\arguments{
\item{x}{(\code{JoinKeys}) object}
}
\value{
number of relationship pairs and primary keys defined in \code{JoinKeys}
object
}
\description{
Length of an Object
}
