% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_dataset.R
\name{load_datasets}
\alias{load_datasets}
\alias{load_datasets.TealDataset}
\alias{load_datasets.TealDatasetConnector}
\alias{load_datasets.TealDataConnector}
\alias{load_datasets.TealData}
\title{Load datasets}
\usage{
load_datasets(x, ...)

\method{load_datasets}{TealDataset}(x, ...)

\method{load_datasets}{TealDatasetConnector}(x, args = NULL, try = FALSE, ...)

\method{load_datasets}{TealDataConnector}(x, ...)

\method{load_datasets}{TealData}(x, ...)
}
\arguments{
\item{x}{(\code{\link{TealData}}, \code{\link{TealDataset}} or \code{\link{TealDatasetConnector}})}

\item{...}{(not used)\cr
only for support of S3}

\item{args}{(\code{NULL} or named \code{list})\cr
additional dynamic arguments passed to function which loads the data. Applicable only on \code{\link{TealDatasetConnector}})}

\item{try}{(\code{logical})\cr
whether perform function evaluation inside \code{try} clause. Applicable only on \code{\link{TealDatasetConnector}})}
}
\value{
If executed in the interactive session shiny app is opened to load the data. If executed in
shiny application - it returns shiny server module.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
\examples{

# TealDataset ------
ADSL <- teal.data::example_cdisc_data("ADSL")
x <- dataset("ADSL", x = ADSL)

load_datasets(x)

# TealDatasetConnector ------
random_data_connector <- function(dataname) {
  fun_dataset_connector(
    dataname = dataname,
    fun = teal.data::example_cdisc_data,
    fun_args = list(dataname = dataname),
  )
}

adsl <- random_data_connector(dataname = "ADSL")
load_datasets(adsl)
get_dataset(adsl)

adae <- random_data_connector(dataname = "ADAE")
load_datasets(adae)
get_dataset(adae)

# TealDataConnector --------
random_data_connector <- function(dataname) {
  fun_dataset_connector(
    dataname = dataname,
    fun = teal.data::example_cdisc_data,
    fun_args = list(dataname = dataname),
  )
}

adsl <- random_data_connector(dataname = "ADSL")
adrs <- random_data_connector(dataname = "ADRS")

rdc <- cdisc_data(adsl, adrs)
\dontrun{
load_datasets(rdc)
}

# TealData --------
random_data_connector <- function(dataname) {
  fun_dataset_connector(
    dataname = dataname,
    fun = teal.data::example_cdisc_data,
    fun_args = list(dataname = dataname),
  )
}

adsl <- random_data_connector(dataname = "ADSL")
adlb <- random_data_connector(dataname = "ADLB")
adrs <- random_data_connector(dataname = "ADRS")

tc <- cdisc_data(adsl, adlb, adrs)
\dontrun{
load_datasets(tc)
}
}
