% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testhat-helpers.R
\name{local_join_keys}
\alias{local_join_keys}
\title{(test helper) Generate a join_keys}
\usage{
local_join_keys(dataset_1 = "ds1", keys = c("id"))
}
\arguments{
\item{dataset_1}{\code{character(1)} name of dataset to add.}

\item{keys}{\code{character(1)} primary key for \code{dataset_1} (optionally named).}
}
\value{
\code{join_keys} object with a primary key
}
\description{
(test helper) Generate a join_keys
}
\keyword{internal}
