% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testhat-helpers.R
\name{local_teal_data}
\alias{local_teal_data}
\title{(test helper) Generate a teal_data dataset with sample data and join_keys}
\usage{
local_teal_data()
}
\value{
\code{teal_data}
}
\description{
(test helper) Generate a teal_data dataset with sample data and join_keys
}
\keyword{internal}
