% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAETealDataset.R
\name{MAETealDataset}
\alias{MAETealDataset}
\title{R6 Class representing a \code{MultiAssayExperiment} object with its attributes}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
Any \code{MultiAssayExperiment} object can be stored inside this \code{MAETealDataset}.
Some attributes like colnames, dimension or column names for a specific type will
be automatically derived.
}
\section{Super class}{
\code{\link[teal.data:TealDataset]{teal.data::TealDataset}} -> \code{MAETealDataset}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-MAETealDataset-new}{\code{MAETealDataset$new()}}
\item \href{#method-MAETealDataset-check}{\code{MAETealDataset$check()}}
\item \href{#method-MAETealDataset-check_keys}{\code{MAETealDataset$check_keys()}}
\item \href{#method-MAETealDataset-get_colnames}{\code{MAETealDataset$get_colnames()}}
\item \href{#method-MAETealDataset-get_column_labels}{\code{MAETealDataset$get_column_labels()}}
\item \href{#method-MAETealDataset-get_ncol}{\code{MAETealDataset$get_ncol()}}
\item \href{#method-MAETealDataset-get_nrow}{\code{MAETealDataset$get_nrow()}}
\item \href{#method-MAETealDataset-get_rownames}{\code{MAETealDataset$get_rownames()}}
\item \href{#method-MAETealDataset-print}{\code{MAETealDataset$print()}}
\item \href{#method-MAETealDataset-clone}{\code{MAETealDataset$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="teal.data" data-topic="TealDataset" data-id="get_attrs"><a href='../../teal.data/html/TealDataset.html#method-TealDataset-get_attrs'><code>teal.data::TealDataset$get_attrs()</code></a></span></li>
<li><span class="pkg-link" data-pkg="teal.data" data-topic="TealDataset" data-id="get_character_colnames"><a href='../../teal.data/html/TealDataset.html#method-TealDataset-get_character_colnames'><code>teal.data::TealDataset$get_character_colnames()</code></a></span></li>
<li><span class="pkg-link" data-pkg="teal.data" data-topic="TealDataset" data-id="get_code"><a href='../../teal.data/html/TealDataset.html#method-TealDataset-get_code'><code>teal.data::TealDataset$get_code()</code></a></span></li>
<li><span class="pkg-link" data-pkg="teal.data" data-topic="TealDataset" data-id="get_code_class"><a href='../../teal.data/html/TealDataset.html#method-TealDataset-get_code_class'><code>teal.data::TealDataset$get_code_class()</code></a></span></li>
<li><span class="pkg-link" data-pkg="teal.data" data-topic="TealDataset" data-id="get_dataname"><a href='../../teal.data/html/TealDataset.html#method-TealDataset-get_dataname'><code>teal.data::TealDataset$get_dataname()</code></a></span></li>
<li><span class="pkg-link" data-pkg="teal.data" data-topic="TealDataset" data-id="get_datanames"><a href='../../teal.data/html/TealDataset.html#method-TealDataset-get_datanames'><code>teal.data::TealDataset$get_datanames()</code></a></span></li>
<li><span class="pkg-link" data-pkg="teal.data" data-topic="TealDataset" data-id="get_dataset"><a href='../../teal.data/html/TealDataset.html#method-TealDataset-get_dataset'><code>teal.data::TealDataset$get_dataset()</code></a></span></li>
<li><span class="pkg-link" data-pkg="teal.data" data-topic="TealDataset" data-id="get_dataset_label"><a href='../../teal.data/html/TealDataset.html#method-TealDataset-get_dataset_label'><code>teal.data::TealDataset$get_dataset_label()</code></a></span></li>
<li><span class="pkg-link" data-pkg="teal.data" data-topic="TealDataset" data-id="get_factor_colnames"><a href='../../teal.data/html/TealDataset.html#method-TealDataset-get_factor_colnames'><code>teal.data::TealDataset$get_factor_colnames()</code></a></span></li>
<li><span class="pkg-link" data-pkg="teal.data" data-topic="TealDataset" data-id="get_join_keys"><a href='../../teal.data/html/TealDataset.html#method-TealDataset-get_join_keys'><code>teal.data::TealDataset$get_join_keys()</code></a></span></li>
<li><span class="pkg-link" data-pkg="teal.data" data-topic="TealDataset" data-id="get_keys"><a href='../../teal.data/html/TealDataset.html#method-TealDataset-get_keys'><code>teal.data::TealDataset$get_keys()</code></a></span></li>
<li><span class="pkg-link" data-pkg="teal.data" data-topic="TealDataset" data-id="get_metadata"><a href='../../teal.data/html/TealDataset.html#method-TealDataset-get_metadata'><code>teal.data::TealDataset$get_metadata()</code></a></span></li>
<li><span class="pkg-link" data-pkg="teal.data" data-topic="TealDataset" data-id="get_mutate_code_class"><a href='../../teal.data/html/TealDataset.html#method-TealDataset-get_mutate_code_class'><code>teal.data::TealDataset$get_mutate_code_class()</code></a></span></li>
<li><span class="pkg-link" data-pkg="teal.data" data-topic="TealDataset" data-id="get_mutate_vars"><a href='../../teal.data/html/TealDataset.html#method-TealDataset-get_mutate_vars'><code>teal.data::TealDataset$get_mutate_vars()</code></a></span></li>
<li><span class="pkg-link" data-pkg="teal.data" data-topic="TealDataset" data-id="get_numeric_colnames"><a href='../../teal.data/html/TealDataset.html#method-TealDataset-get_numeric_colnames'><code>teal.data::TealDataset$get_numeric_colnames()</code></a></span></li>
<li><span class="pkg-link" data-pkg="teal.data" data-topic="TealDataset" data-id="get_raw_data"><a href='../../teal.data/html/TealDataset.html#method-TealDataset-get_raw_data'><code>teal.data::TealDataset$get_raw_data()</code></a></span></li>
<li><span class="pkg-link" data-pkg="teal.data" data-topic="TealDataset" data-id="get_row_labels"><a href='../../teal.data/html/TealDataset.html#method-TealDataset-get_row_labels'><code>teal.data::TealDataset$get_row_labels()</code></a></span></li>
<li><span class="pkg-link" data-pkg="teal.data" data-topic="TealDataset" data-id="get_var_r6"><a href='../../teal.data/html/TealDataset.html#method-TealDataset-get_var_r6'><code>teal.data::TealDataset$get_var_r6()</code></a></span></li>
<li><span class="pkg-link" data-pkg="teal.data" data-topic="TealDataset" data-id="get_vars"><a href='../../teal.data/html/TealDataset.html#method-TealDataset-get_vars'><code>teal.data::TealDataset$get_vars()</code></a></span></li>
<li><span class="pkg-link" data-pkg="teal.data" data-topic="TealDataset" data-id="is_mutate_delayed"><a href='../../teal.data/html/TealDataset.html#method-TealDataset-is_mutate_delayed'><code>teal.data::TealDataset$is_mutate_delayed()</code></a></span></li>
<li><span class="pkg-link" data-pkg="teal.data" data-topic="TealDataset" data-id="is_pulled"><a href='../../teal.data/html/TealDataset.html#method-TealDataset-is_pulled'><code>teal.data::TealDataset$is_pulled()</code></a></span></li>
<li><span class="pkg-link" data-pkg="teal.data" data-topic="TealDataset" data-id="merge_join_keys"><a href='../../teal.data/html/TealDataset.html#method-TealDataset-merge_join_keys'><code>teal.data::TealDataset$merge_join_keys()</code></a></span></li>
<li><span class="pkg-link" data-pkg="teal.data" data-topic="TealDataset" data-id="mutate"><a href='../../teal.data/html/TealDataset.html#method-TealDataset-mutate'><code>teal.data::TealDataset$mutate()</code></a></span></li>
<li><span class="pkg-link" data-pkg="teal.data" data-topic="TealDataset" data-id="mutate_join_keys"><a href='../../teal.data/html/TealDataset.html#method-TealDataset-mutate_join_keys'><code>teal.data::TealDataset$mutate_join_keys()</code></a></span></li>
<li><span class="pkg-link" data-pkg="teal.data" data-topic="TealDataset" data-id="reassign_datasets_vars"><a href='../../teal.data/html/TealDataset.html#method-TealDataset-reassign_datasets_vars'><code>teal.data::TealDataset$reassign_datasets_vars()</code></a></span></li>
<li><span class="pkg-link" data-pkg="teal.data" data-topic="TealDataset" data-id="recreate"><a href='../../teal.data/html/TealDataset.html#method-TealDataset-recreate'><code>teal.data::TealDataset$recreate()</code></a></span></li>
<li><span class="pkg-link" data-pkg="teal.data" data-topic="TealDataset" data-id="set_code"><a href='../../teal.data/html/TealDataset.html#method-TealDataset-set_code'><code>teal.data::TealDataset$set_code()</code></a></span></li>
<li><span class="pkg-link" data-pkg="teal.data" data-topic="TealDataset" data-id="set_dataset_label"><a href='../../teal.data/html/TealDataset.html#method-TealDataset-set_dataset_label'><code>teal.data::TealDataset$set_dataset_label()</code></a></span></li>
<li><span class="pkg-link" data-pkg="teal.data" data-topic="TealDataset" data-id="set_join_keys"><a href='../../teal.data/html/TealDataset.html#method-TealDataset-set_join_keys'><code>teal.data::TealDataset$set_join_keys()</code></a></span></li>
<li><span class="pkg-link" data-pkg="teal.data" data-topic="TealDataset" data-id="set_keys"><a href='../../teal.data/html/TealDataset.html#method-TealDataset-set_keys'><code>teal.data::TealDataset$set_keys()</code></a></span></li>
<li><span class="pkg-link" data-pkg="teal.data" data-topic="TealDataset" data-id="set_vars"><a href='../../teal.data/html/TealDataset.html#method-TealDataset-set_vars'><code>teal.data::TealDataset$set_vars()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MAETealDataset-new"></a>}}
\if{latex}{\out{\hypertarget{method-MAETealDataset-new}{}}}
\subsection{Method \code{new()}}{
Create a new object of \code{MAETealDataset} class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MAETealDataset$new(
  dataname,
  x,
  keys = character(0),
  code = character(0),
  label = character(0),
  vars = list(),
  metadata = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dataname}}{(\code{character})\cr
A given name for the dataset it may not contain spaces}

\item{\code{x}}{(\code{MultiAssayExperiment})\cr}

\item{\code{keys}}{optional, (\code{character})\cr
A vector of primary keys}

\item{\code{code}}{(\code{character} or \code{CodeClass})\cr
A character string defining the code needed to produce the data set in \code{x}.
initialize()\code{and}recreate()\verb{accept code as}CodeClass` also
which is needed to preserve the code uniqueness and correct order.}

\item{\code{label}}{(\code{character})\cr
Label to describe the dataset}

\item{\code{vars}}{(named \code{list})) \cr
In case when this object code depends on other \code{TealDataset} object(s) or
other constant value, this/these object(s) should be included as named
element(s) of the list. For example if this object code needs \code{ADSL}
object we should specify \verb{vars = list(ADSL = <adsl object>)}.
It's recommended to include \code{TealDataset} or \code{TealDatasetConnector} objects to
the \code{vars} list to preserve reproducibility. Please note that \code{vars}
are included to this object as local \code{vars} and they cannot be modified
within another dataset.}

\item{\code{metadata}}{(named \code{list} or \code{NULL}) \cr
Field containing metadata about the dataset. Each element of the list
should be atomic and length one.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MAETealDataset-check"></a>}}
\if{latex}{\out{\hypertarget{method-MAETealDataset-check}{}}}
\subsection{Method \code{check()}}{
Check to determine if the raw data is reproducible from the \code{get_code()} code.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MAETealDataset$check()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{TRUE} if the dataset generated from evaluating the
\code{get_code()} code is identical to the raw data, else \code{FALSE}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MAETealDataset-check_keys"></a>}}
\if{latex}{\out{\hypertarget{method-MAETealDataset-check_keys}{}}}
\subsection{Method \code{check_keys()}}{
Check if keys has been specified correctly for dataset. Set of \code{keys}
should distinguish unique rows or be \code{character(0)}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MAETealDataset$check_keys(keys = private$.keys)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{keys}}{optional, (\code{character})\cr
A vector of primary keys}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{TRUE} if dataset has been already pulled, else \code{FALSE}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MAETealDataset-get_colnames"></a>}}
\if{latex}{\out{\hypertarget{method-MAETealDataset-get_colnames}{}}}
\subsection{Method \code{get_colnames()}}{
Derive the column names
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MAETealDataset$get_colnames()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{character} vector.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MAETealDataset-get_column_labels"></a>}}
\if{latex}{\out{\hypertarget{method-MAETealDataset-get_column_labels}{}}}
\subsection{Method \code{get_column_labels()}}{
Derive the column labels
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MAETealDataset$get_column_labels()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{character} vector.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MAETealDataset-get_ncol"></a>}}
\if{latex}{\out{\hypertarget{method-MAETealDataset-get_ncol}{}}}
\subsection{Method \code{get_ncol()}}{
Get the number of columns of the data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MAETealDataset$get_ncol()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{numeric} vector
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MAETealDataset-get_nrow"></a>}}
\if{latex}{\out{\hypertarget{method-MAETealDataset-get_nrow}{}}}
\subsection{Method \code{get_nrow()}}{
Get the number of rows of the data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MAETealDataset$get_nrow()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{numeric} vector
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MAETealDataset-get_rownames"></a>}}
\if{latex}{\out{\hypertarget{method-MAETealDataset-get_rownames}{}}}
\subsection{Method \code{get_rownames()}}{
Derive the row names
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MAETealDataset$get_rownames()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{character} vector.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MAETealDataset-print"></a>}}
\if{latex}{\out{\hypertarget{method-MAETealDataset-print}{}}}
\subsection{Method \code{print()}}{
Prints this \code{MAETealDataset}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MAETealDataset$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{additional arguments to the printing method}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
invisibly self
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-MAETealDataset-clone"></a>}}
\if{latex}{\out{\hypertarget{method-MAETealDataset-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{MAETealDataset$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
