% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_keys.R
\name{merge_join_keys}
\alias{merge_join_keys}
\alias{merge_join_keys.default}
\alias{merge_join_keys.join_keys}
\title{Merging a list (or one) of \code{join_keys} objects into the current \code{join_keys} object}
\usage{
merge_join_keys(join_keys_obj, new_join_keys)

\method{merge_join_keys}{default}(join_keys_obj, new_join_keys)

\method{merge_join_keys}{join_keys}(join_keys_obj, new_join_keys)
}
\arguments{
\item{join_keys_obj}{(\code{join_keys}) object to merge the new_join_keys.}

\item{new_join_keys}{\code{list} of \code{join_keys} objects or single \code{join_keys} object}
}
\value{
a new \code{join_keys} object with the resulting merge.
}
\description{
Merging a list (or one) of \code{join_keys} objects into the current \code{join_keys} object
}
\keyword{internal}
