% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_keys.R
\name{names<-.join_keys}
\alias{names<-.join_keys}
\title{The Names of an \code{join_keys} Object}
\usage{
\method{names}{join_keys}(x) <- value
}
\arguments{
\item{x}{an \R object.}

\item{value}{a character vector of up to the same length as \code{x}, or
    \code{NULL}.}
}
\description{
The Names of an \code{join_keys} Object
}
