% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teal_data-names.R
\name{names.teal_data}
\alias{names.teal_data}
\title{Names of data sets in \code{teal_data} object}
\usage{
\method{names}{teal_data}(x)
}
\arguments{
\item{x}{A (\code{teal_data}) object to access or modify.}
}
\value{
A character vector of names.
}
\description{
Functions to get the names of a \code{teal_data} object.
The names are extrapolated from the objects in the \code{qenv} environment and
are not stored statically, unlike the normal behavior of \code{names()} function.
}
\details{
Objects named with a \code{.} (dot) prefix will be ignored and not returned,
unless \code{all.names} parameter is set to \code{TRUE}.
}
\examples{
td <- teal_data(iris = iris)
td <- within(td, mtcars <- mtcars)
names(td)

td <- within(td, .CO2 <- CO2)
names(td)

}
