% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdata.R
\name{new_tdata}
\alias{new_tdata}
\alias{new_tdata,list,expression-method}
\alias{new_tdata,list,language-method}
\alias{new_tdata,list,character-method}
\title{Initialize \code{tdata} object}
\usage{
new_tdata(env = new.env(), code = expression(), keys = join_keys())

\S4method{new_tdata}{list,expression}(env = new.env(), code = expression(), keys = join_keys())

\S4method{new_tdata}{list,language}(env = new.env(), code = expression(), keys = join_keys())

\S4method{new_tdata}{list,character}(env = new.env(), code = expression(), keys = join_keys())
}
\arguments{
\item{env}{(\code{list}) List of data.}

\item{code}{(\code{character(1)} or \code{language}) code to evaluate. Accepts and stores comments also.}
}
\description{
Initialize \code{tdata} object.
}
\examples{
new_tdata(env = list(a = 1), code = quote(a <- 1))
new_tdata(env = list(a = 1), code = parse(text = "a <- 1"))
new_tdatas(env = list(a = 1), code = "a <- 1")

}
