% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdata2-constructors.r
\name{new_tdata2}
\alias{new_tdata2}
\alias{new_tdata2,list,expression-method}
\alias{new_tdata2,list,language-method}
\alias{new_tdata2,list,character-method}
\title{Initialize \code{tdata} object}
\usage{
new_tdata2(env = new.env(), code = expression(), join_keys = join_keys())

\S4method{new_tdata2}{list,expression}(env, code)

\S4method{new_tdata2}{list,language}(env, code)

\S4method{new_tdata2}{list,character}(env = new.env(), code = expression(), join_keys = join_keys())
}
\arguments{
\item{env}{(\code{list}) List of data.}

\item{code}{(\code{character(1)} or \code{language}) code to evaluate. Accepts and stores comments also.}
}
\description{
Initialize \code{tdata} object.
}
\examples{
new_tdata2(env = list(a = 1), code = quote(a <- 1))
new_tdata2(env = list(a = 1), code = parse(text = "a <- 1"))
new_tdatas(env = list(a = 1), code = "a <- 1")

}
