% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_keys-parents.R
\name{parent}
\alias{parent}
\title{Getter and setter for specific parent}
\usage{
parent(x, dataset_name)
}
\arguments{
\item{x}{(\code{join_keys}) object to retrieve.}

\item{dataset_name}{(\code{character(1)})}
}
\description{
Getter and setter for specific parent
}
\examples{
jk <- join_keys(join_key("ds1", "ds2", "key"))
parent(jk, "ds2")
parents(jk) <- list("ds2" = "ds1")
parent(jk, "ds2")
}
