% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_keys.R
\name{print.join_keys}
\alias{print.join_keys}
\title{Prints \code{join_keys}.}
\usage{
\method{print}{join_keys}(x, ...)
}
\arguments{
\item{x}{an object used to select a method.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
the \code{x} parameter
}
\description{
Prints \code{join_keys}.
}
