% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_keys.R
\name{print.JoinKeys}
\alias{print.JoinKeys}
\title{Prints \code{JoinKeys}.}
\usage{
\method{print}{JoinKeys}(x, ...)
}
\arguments{
\item{x}{an object used to select a method.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
the \code{x} parameter
}
\description{
Prints \code{JoinKeys}.
}
