% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CallablePythonCode.R
\name{python_code}
\alias{python_code}
\title{Python Code}
\usage{
python_code(code = character(0), script = character(0))
}
\arguments{
\item{code}{(\code{character})\cr
Code to mutate the dataset. Must contain the \code{dataset$dataname}.}

\item{script}{(\code{character})\cr
file that contains python Code that can be read using \code{reticulate::py_run_script}.}
}
\value{
(\code{PythonCodeClass}) object containing python code
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
Create a python code object directly from python code or a
script containing python code.
}
\details{
Used to mutate dataset connector objects with python code. See
\code{\link{mutate_dataset}} or \code{\link{mutate_data}} for details.
}
\examples{
\dontrun{
library(reticulate)
library(magrittr)

# mutate dataset object

ADSL <- teal.data::rADSL

x <- dataset_connector(dataname = "ADSL",
                       pull_callable = callable_function(fun = function(ADSL) ADSL <- teal.data::rADSL),
                       keys = get_cdisc_keys("ADSL"),
                       label = "Example label")

x \%>\% mutate_dataset(python_code("import pandas as pd
r.ADSL = pd.DataFrame({'x': [1]})"))

x$get_code()
x$pull()
x$get_raw_data()

# mutate data object

y <- 8
ADLB <- teal.data::rADLB
adsl_connector <- dataset_connector(dataname = "ADSL",
                                    pull_callable = callable_function(fun = function(ADSL) ADSL <- teal.data::rADSL) \%>\% # nolint
                                      set_args(list(ADSL = as.name("ADSL"))),,
                                    keys = get_cdisc_keys("ADSL"),
                                    label = "ADSL connector")
adlb_connector <- dataset_connector(dataname = "ADLB",
                                    pull_callable = callable_function(fun = function(ADLB) ADLB <- teal.data::rADLB) \%>\% # nolint
                                      set_args(list(ADLB = as.name("ADLB"))),
                                    keys = get_cdisc_keys("ADLB"),
                                    label = "ADLB connector")
tc <- cdisc_data(
  adsl_connector,
  adlb_connector
)

tc \%>\% mutate_data(python_code("import pandas as pd
r.ADSL = pd.DataFrame({'STUDYID': [r.y], 'USUBJID': [r.y]})"), vars = list(y = y))


load_datasets(tc) # submit all
ds <- tc$get_dataset("ADSL")
ds$get_raw_data()
}
}
