% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CallablePythonCode.R
\name{PythonCodeClass}
\alias{PythonCodeClass}
\title{A \code{CallablePythonCode} class of objects}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\section{Super class}{
\code{\link[teal.data:CodeClass]{teal.data::CodeClass}} -> \code{PythonCodeClass}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-PythonCodeClass-eval}{\code{PythonCodeClass$eval()}}
\item \href{#method-PythonCodeClass-clone}{\code{PythonCodeClass$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="teal.data" data-topic="CodeClass" data-id="append"><a href='../../teal.data/html/CodeClass.html#method-CodeClass-append'><code>teal.data::CodeClass$append()</code></a></span></li>
<li><span class="pkg-link" data-pkg="teal.data" data-topic="CodeClass" data-id="get_code"><a href='../../teal.data/html/CodeClass.html#method-CodeClass-get_code'><code>teal.data::CodeClass$get_code()</code></a></span></li>
<li><span class="pkg-link" data-pkg="teal.data" data-topic="CodeClass" data-id="initialize"><a href='../../teal.data/html/CodeClass.html#method-CodeClass-initialize'><code>teal.data::CodeClass$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="teal.data" data-topic="CodeClass" data-id="set_code"><a href='../../teal.data/html/CodeClass.html#method-CodeClass-set_code'><code>teal.data::CodeClass$set_code()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PythonCodeClass-eval"></a>}}
\if{latex}{\out{\hypertarget{method-PythonCodeClass-eval}{}}}
\subsection{Method \code{eval()}}{
Evaluates internal code within environment
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PythonCodeClass$eval(
  vars = list(),
  dataname = NULL,
  envir = new.env(parent = parent.env(.GlobalEnv))
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{vars}}{(named \code{list}) additional pre-requisite vars to execute code}

\item{\code{dataname}}{(\code{character}) name of the data frame object to be returned}

\item{\code{envir}}{(\code{environment}) environment in which code will be evaluated}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{data.frame} containing the mutated dataset
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-PythonCodeClass-clone"></a>}}
\if{latex}{\out{\hypertarget{method-PythonCodeClass-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PythonCodeClass$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
