% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rADEX}
\alias{rADEX}
\title{Random response analysis dataset}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 6400 rows and 79 columns.
}
\source{
\link[random.cdisc.data:cadex]{random.cdisc.data::cadex}
}
\usage{
rADEX
}
\description{
Random response analysis dataset
}
\keyword{datasets}
