% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rADLB}
\alias{rADLB}
\title{Random laboratory data analysis dataset}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 8400 rows and 102 columns.
}
\source{
\link[random.cdisc.data:cadlb]{random.cdisc.data::cadlb}
}
\usage{
rADLB
}
\description{
Random laboratory data analysis dataset
}
\keyword{datasets}
