% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{random_cdisc_data}
\alias{random_cdisc_data}
\alias{rADSL}
\alias{rADAE}
\alias{rADCM}
\alias{rADEX}
\alias{rADLB}
\alias{rADRS}
\alias{rADTR}
\alias{rADTTE}
\alias{rADVS}
\title{Simulated \code{CDISC} Data for Examples}
\format{
\itemize{
\item \code{rADSL}: An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 400 rows and 55 columns.
}

\itemize{
\item \code{rADAE}: An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1934 rows and 92 columns.
}

\itemize{
\item \code{rADCM}: An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 3685 rows and 83 columns.
}

\itemize{
\item \code{rADEX}: An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 6400 rows and 79 columns.
}

\itemize{
\item \code{rADLB}: An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 8400 rows and 102 columns.
}

\itemize{
\item \code{rADRS}: An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 3200 rows and 65 columns.
}

\itemize{
\item \code{rADTR}: An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2800 rows and 76 columns.
}

\itemize{
\item \code{rADTTE}: An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2000 rows and 67 columns.
}

\itemize{
\item \code{rADVS}: An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 16800 rows and 87 columns.
}
}
\source{
\itemize{
\item \code{rADSL}: \link[random.cdisc.data:cadsl]{random.cdisc.data::cadsl}
}

\itemize{
\item \code{rADAE}: \link[random.cdisc.data:cadae]{random.cdisc.data::cadae}
}

\itemize{
\item \code{rADCM}: \link[random.cdisc.data:cadcm]{random.cdisc.data::cadcm}
}

\itemize{
\item \code{rADEX}: \link[random.cdisc.data:cadex]{random.cdisc.data::cadex}
}

\itemize{
\item \code{rADLB}: \link[random.cdisc.data:cadlb]{random.cdisc.data::cadlb}
}

\itemize{
\item \code{rADRS}: \link[random.cdisc.data:cadrs]{random.cdisc.data::cadrs}
}

\itemize{
\item \code{rADTR}: \link[random.cdisc.data:cadtr]{random.cdisc.data::cadtr}
}

\itemize{
\item \code{rADTTE}: \link[random.cdisc.data:cadtte]{random.cdisc.data::cadtte}
}

\itemize{
\item \code{rADVS}: \link[random.cdisc.data:cadvs]{random.cdisc.data::cadvs}
}
}
\usage{
rADSL

rADAE

rADCM

rADEX

rADLB

rADRS

rADTR

rADTTE

rADVS
}
\description{
This package contains random datasets that are used for testing purposes in the \code{teal.data} package:
\itemize{
\item \code{rADSL} Random subject-level analysis dataset
}

\itemize{
\item \code{rADAE}: Random adverse events analysis dataset
}

\itemize{
\item \code{rADCM}: Random concomitant medications analysis dataset
}

\itemize{
\item \code{rADEX}: Random response analysis dataset
}

\itemize{
\item \code{rADLB}: Random laboratory data analysis dataset
}

\itemize{
\item \code{rADRS}: Random response analysis dataset
}

\itemize{
\item \code{rADTR}: Random tumor response analysis dataset
}

\itemize{
\item \code{rADTTE}: Random time to event analysis dataset
}

\itemize{
\item \code{rADVS}: Random vital signs analysis dataset
}
}
\keyword{datasets}
