% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rADTR}
\alias{rADTR}
\title{Random data \code{rADTR}}
\format{
An object of class \code{data.frame} with 2800 rows and 76 columns.
}
\source{
internal
}
\usage{
rADTR
}
\description{
Random data \code{rADTR}
}
\keyword{datasets}
\keyword{internal}
