% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{random_cdisc_data}
\alias{random_cdisc_data}
\alias{rADSL}
\title{Simulated \code{CDISC} Data for Examples}
\format{
\itemize{
\item \code{rADSL}: An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 400 rows and 55 columns.
}
}
\source{
\itemize{
\item \code{rADSL}: \link[random.cdisc.data:cadsl]{random.cdisc.data::cadsl}
}
}
\usage{
rADSL
}
\description{
This package contains random datasets that are used for testing purposes in the \code{teal.data} package:
\itemize{
\item \code{rADSL} Random subject-level analysis dataset
}
}
\keyword{datasets}
