% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TealDataConnector.R
\name{relational_data_connector}
\alias{relational_data_connector}
\title{The constructor for \code{TealDataConnector} class.}
\usage{
relational_data_connector(connection, connectors)
}
\arguments{
\item{connection}{(\code{TealDataConnection})\cr
connection to data source}

\item{connectors}{(\code{list} of \code{TealDatasetConnector} elements)\cr
list with dataset connectors}
}
\value{
\code{TealDataConnector} object
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
\examples{
library(magrittr)
pull_adsl <- function(ADSL, n) ADSL <- head(teal.data::rADSL, n)
adsl_connector <- dataset_connector(
  dataname = "ADSL",
  pull_callable = callable_function(fun = pull_adsl) \%>\% # nolint
    set_args(list(ADSL = as.name("ADSL"))),
  keys = get_cdisc_keys("ADSL"),
  label = "ADSL connector"
)

pull_adlb <- function(ADLB, n) ADLB <- head(teal.data::rADLB, n)
adlb_connector <- dataset_connector(
  dataname = "ADLB",
  pull_callable = callable_function(fun = pull_adlb) \%>\% # nolint
    set_args(list(ADLB = as.name("ADLB"))),
  keys = get_cdisc_keys("ADLB"),
  label = "ADLB connector"
)

open_fun <- callable_function(library)
open_fun$set_args(list(package = "teal.data"))

con <- data_connection(open_fun = open_fun)
con$set_open_server(
  function(id, connection) {
    moduleServer(
      id = id,
      module = function(input, output, session) {
        connection$open(try = TRUE)
        return(invisible(connection))
      }
    )
  }
)

x <- relational_data_connector(connection = con, connectors = list(adsl_connector, adlb_connector))

x$set_ui(
  function(id, connection, connectors) {
    ns <- NS(id)
    tagList(
      connection$get_open_ui(ns("open_connection")),
      numericInput(inputId = ns("n"), label = "Choose number of records", min = 0, value = 1),
      do.call(
        what = "tagList",
        args = lapply(
          connectors,
          function(connector) {
            div(
              connector$get_ui(
                id = ns(connector$get_dataname())
              ),
              br()
            )
          }
        )
      )
    )
  }
)

x$set_server(
  function(id, connection, connectors) {
    moduleServer(
      id = id,
      module = function(input, output, session) {
        # opens connection
        connection$get_open_server()(id = "open_connection", connection = connection)
        if (connection$is_opened()) {
          for (connector in connectors) {
            set_args(connector, args = list(n = input$n))
            # pull each dataset
            connector$get_server()(id = connector$get_dataname())
            if (connector$is_failed()) {
              break
            }
          }
        }
      }
    )
  }
)
\dontrun{
x$launch()
x$get_datasets()
}

}
