% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teal_data-show.R
\name{show,teal_data-method}
\alias{show,teal_data-method}
\title{Show \code{teal_data} object}
\usage{
\S4method{show}{teal_data}(object)
}
\arguments{
\item{object}{(\code{teal_data})}
}
\value{
Input \code{teal_data} object.
}
\description{
Prints \code{teal_data} object.
}
\examples{
teal_data()
teal_data(x = iris, code = "x = iris")
verify(teal_data(x = iris, code = "x = iris"))
}
