% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JoinKeys.R
\name{[.JoinKeys}
\alias{[.JoinKeys}
\alias{[<-.JoinKeys}
\title{Getter for JoinKeys that returns the relationship between pairs of datasets}
\usage{
\method{[}{JoinKeys}(x, dataset_1, dataset_2 = NULL)

\method{[}{JoinKeys}(x, dataset_1, dataset_2 = NULL) <- value
}
\description{
Getter for JoinKeys that returns the relationship between pairs of datasets
}
