% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_join_keys.R
\name{[.Placeholder}
\alias{[.Placeholder}
\alias{[<-.Placeholder}
\title{Getter for JoinKeys that returns the relationship between pairs of datasets}
\usage{
\method{[}{Placeholder}(x, dataset_1, dataset_2 = NULL)

\method{[}{Placeholder}(data, dataset_1, dataset_2 = NULL) <- value
}
\arguments{
\item{x}{JoinKeys object to extract the join keys}

\item{dataset_1}{(\code{character}) name of first dataset.}

\item{dataset_2}{(\code{character}) name of second dataset.}

\item{keys}{value to assign}
}
\description{
Getter for JoinKeys that returns the relationship between pairs of datasets
}
\examples{
jk <- new_join_keys()
join_keys(jk) <- list(ds1 = list(ds2 = "some_col"))
jk["ds1", "ds2"]
jk <- new_join_keys()
jk["ds1", "ds2"]
join_keys(jk) <- join_key("ds1", "ds2", "some_col")
jk["ds1", "ds2"]
jk["ds1", "ds2"] <- "new_col"
jk["ds1", "ds2"]
}
\keyword{internal}
