% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TealDataAbstract.R
\name{TealDataAbstract}
\alias{TealDataAbstract}
\title{\code{TealDataAbstract} class}
\description{
Abstract class containing code for handling set of datasets.
}
\keyword{internal}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{id}}{String used to create unique GUI elements}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TealDataAbstract-new}{\code{TealDataAbstract$new()}}
\item \href{#method-TealDataAbstract-check}{\code{TealDataAbstract$check()}}
\item \href{#method-TealDataAbstract-check_reproducibility}{\code{TealDataAbstract$check_reproducibility()}}
\item \href{#method-TealDataAbstract-execute_mutate}{\code{TealDataAbstract$execute_mutate()}}
\item \href{#method-TealDataAbstract-get_check_result}{\code{TealDataAbstract$get_check_result()}}
\item \href{#method-TealDataAbstract-get_code}{\code{TealDataAbstract$get_code()}}
\item \href{#method-TealDataAbstract-get_code_class}{\code{TealDataAbstract$get_code_class()}}
\item \href{#method-TealDataAbstract-get_datanames}{\code{TealDataAbstract$get_datanames()}}
\item \href{#method-TealDataAbstract-get_dataset}{\code{TealDataAbstract$get_dataset()}}
\item \href{#method-TealDataAbstract-get_datasets}{\code{TealDataAbstract$get_datasets()}}
\item \href{#method-TealDataAbstract-get_items}{\code{TealDataAbstract$get_items()}}
\item \href{#method-TealDataAbstract-get_check}{\code{TealDataAbstract$get_check()}}
\item \href{#method-TealDataAbstract-is_pulled}{\code{TealDataAbstract$is_pulled()}}
\item \href{#method-TealDataAbstract-mutate}{\code{TealDataAbstract$mutate()}}
\item \href{#method-TealDataAbstract-mutate_dataset}{\code{TealDataAbstract$mutate_dataset()}}
\item \href{#method-TealDataAbstract-set_check}{\code{TealDataAbstract$set_check()}}
\item \href{#method-TealDataAbstract-set_pull_code}{\code{TealDataAbstract$set_pull_code()}}
\item \href{#method-TealDataAbstract-reassign_datasets_vars}{\code{TealDataAbstract$reassign_datasets_vars()}}
\item \href{#method-TealDataAbstract-clone}{\code{TealDataAbstract$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TealDataAbstract-new"></a>}}
\if{latex}{\out{\hypertarget{method-TealDataAbstract-new}{}}}
\subsection{Method \code{new()}}{
Cannot create a \code{TealDataAbstract} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TealDataAbstract$new()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
throws error
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TealDataAbstract-check"></a>}}
\if{latex}{\out{\hypertarget{method-TealDataAbstract-check}{}}}
\subsection{Method \code{check()}}{
Check if the object raw data is reproducible from the \code{get_code()} code.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TealDataAbstract$check()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{NULL} if check step has been disabled
\code{TRUE} if all the datasets generated from evaluating the
\code{get_code()} code are identical to the raw data, else \code{FALSE}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TealDataAbstract-check_reproducibility"></a>}}
\if{latex}{\out{\hypertarget{method-TealDataAbstract-check_reproducibility}{}}}
\subsection{Method \code{check_reproducibility()}}{
Execute \code{check()} and raise an error if it's not reproducible.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TealDataAbstract$check_reproducibility()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
error if code is not reproducible else invisibly nothing
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TealDataAbstract-execute_mutate"></a>}}
\if{latex}{\out{\hypertarget{method-TealDataAbstract-execute_mutate}{}}}
\subsection{Method \code{execute_mutate()}}{
Execute mutate code. Using \verb{mutate_data(set).TealDataAbstract}
does not cause instant execution, the \code{mutate_code} is
delayed and can be evaluated using this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TealDataAbstract$execute_mutate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TealDataAbstract-get_check_result"></a>}}
\if{latex}{\out{\hypertarget{method-TealDataAbstract-get_check_result}{}}}
\subsection{Method \code{get_check_result()}}{
Get result of reproducibility check
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TealDataAbstract$get_check_result()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{NULL} if check has not been called yet, \code{TRUE} / \code{FALSE} otherwise
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TealDataAbstract-get_code"></a>}}
\if{latex}{\out{\hypertarget{method-TealDataAbstract-get_code}{}}}
\subsection{Method \code{get_code()}}{
Get code for all datasets.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TealDataAbstract$get_code(dataname = NULL, deparse = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dataname}}{(\code{character}) dataname or \code{NULL} for all datasets}

\item{\code{deparse}}{(\code{logical}) whether to return the deparsed form of a call}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
(\code{character}) vector of code to generate datasets.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TealDataAbstract-get_code_class"></a>}}
\if{latex}{\out{\hypertarget{method-TealDataAbstract-get_code_class}{}}}
\subsection{Method \code{get_code_class()}}{
Get internal \code{CodeClass} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TealDataAbstract$get_code_class(only_pull = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{only_pull}}{(\code{logical} value)\cr
if \code{TRUE} only code to pull datasets will be returned without the mutate code.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{CodeClass}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TealDataAbstract-get_datanames"></a>}}
\if{latex}{\out{\hypertarget{method-TealDataAbstract-get_datanames}{}}}
\subsection{Method \code{get_datanames()}}{
Get names of the datasets.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TealDataAbstract$get_datanames()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{character} vector with names of all datasets.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TealDataAbstract-get_dataset"></a>}}
\if{latex}{\out{\hypertarget{method-TealDataAbstract-get_dataset}{}}}
\subsection{Method \code{get_dataset()}}{
Get \code{TealDataset} object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TealDataAbstract$get_dataset(dataname = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dataname}}{(\code{character} value)\cr
name of dataset to be returned. If \code{NULL}, all datasets are returned.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{TealDataset}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TealDataAbstract-get_datasets"></a>}}
\if{latex}{\out{\hypertarget{method-TealDataAbstract-get_datasets}{}}}
\subsection{Method \code{get_datasets()}}{
Get \code{list} of \code{TealDataset} objects.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TealDataAbstract$get_datasets()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{list} of \code{TealDataset}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TealDataAbstract-get_items"></a>}}
\if{latex}{\out{\hypertarget{method-TealDataAbstract-get_items}{}}}
\subsection{Method \code{get_items()}}{
Get all datasets and all dataset connectors
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TealDataAbstract$get_items(dataname = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dataname}}{(\code{character} value)\cr
name of dataset connector to be returned. If \code{NULL}, all connectors are returned.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{list} with all datasets and all connectors
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TealDataAbstract-get_check"></a>}}
\if{latex}{\out{\hypertarget{method-TealDataAbstract-get_check}{}}}
\subsection{Method \code{get_check()}}{
Has this data been or will this data be subjected to a reproducibility check
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TealDataAbstract$get_check()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{logical}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TealDataAbstract-is_pulled"></a>}}
\if{latex}{\out{\hypertarget{method-TealDataAbstract-is_pulled}{}}}
\subsection{Method \code{is_pulled()}}{
Check if dataset has already been pulled.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TealDataAbstract$is_pulled()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{TRUE} if dataset has been already pulled, else \code{FALSE}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TealDataAbstract-mutate"></a>}}
\if{latex}{\out{\hypertarget{method-TealDataAbstract-mutate}{}}}
\subsection{Method \code{mutate()}}{
Mutate data by code. Code used in this mutation is not linked to particular
but refers to all datasets.
Consequence of this is that when using \verb{get_code(<dataset>)} this
part of the code will be returned for each specified dataset. This method
should be used only if particular call involve changing multiple datasets.
Otherwise please use \code{mutate_dataset}.
Execution of \code{mutate_code} is delayed after datasets are pulled
(\code{isTRUE(is_pulled)}).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TealDataAbstract$mutate(code, vars = list())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{code}}{(\code{character}) Code to mutate the dataset. Must contain the
\code{dataset$dataname}}

\item{\code{vars}}{(named \code{list})) \cr
In case when this object code depends on other \code{TealDataset} object(s) or
other constant value, this/these object(s) should be included as named
element(s) of the list. For example if this object code needs \code{ADSL}
object we should specify \verb{vars = list(ADSL = <adsl object>)}.
It's recommended to include \code{TealDataset} or \code{TealDatasetConnector} objects to
the \code{vars} list to preserve reproducibility. Please note that \code{vars}
are included to this object as local \code{vars} and they cannot be modified
within another dataset.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
self invisibly for chaining
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TealDataAbstract-mutate_dataset"></a>}}
\if{latex}{\out{\hypertarget{method-TealDataAbstract-mutate_dataset}{}}}
\subsection{Method \code{mutate_dataset()}}{
Mutate dataset by code.
Execution of \code{mutate_code} is delayed after datasets are pulled
(\code{isTRUE(is_pulled)}).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TealDataAbstract$mutate_dataset(dataname, code, vars = list())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dataname}}{(\code{character}) Dataname to be mutated}

\item{\code{code}}{(\code{character}) Code to mutate the dataset. Must contain the
\code{dataset$dataname}}

\item{\code{vars}}{(named \code{list})) \cr
In case when this object code depends on other \code{TealDataset} object(s) or
other constant value, this/these object(s) should be included as named
element(s) of the list. For example if this object code needs \code{ADSL}
object we should specify \verb{vars = list(ADSL = <adsl object>)}.
It's recommended to include \code{TealDataset} or \code{TealDatasetConnector} objects to
the \code{vars} list to preserve reproducibility. Please note that \code{vars}
are included to this object as local \code{vars} and they cannot be modified
within another dataset.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
self invisibly for chaining
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TealDataAbstract-set_check"></a>}}
\if{latex}{\out{\hypertarget{method-TealDataAbstract-set_check}{}}}
\subsection{Method \code{set_check()}}{
Set reproducibility check
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TealDataAbstract$set_check(check = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{check}}{(\code{logical}) whether to perform reproducibility check.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
(\code{self}) invisibly for chaining.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TealDataAbstract-set_pull_code"></a>}}
\if{latex}{\out{\hypertarget{method-TealDataAbstract-set_pull_code}{}}}
\subsection{Method \code{set_pull_code()}}{
Set pull code
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TealDataAbstract$set_pull_code(code)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{code}}{(\code{character} value)\cr
code to reproduce \code{data} in \code{TealDataset} objects. Can't be set if any dataset
has \code{code} set already.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
(\code{self}) invisibly for chaining.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TealDataAbstract-reassign_datasets_vars"></a>}}
\if{latex}{\out{\hypertarget{method-TealDataAbstract-reassign_datasets_vars}{}}}
\subsection{Method \code{reassign_datasets_vars()}}{
Reassign \code{vars} in \code{TealDataset} and \code{TealDatasetConnector} objects
to keep the valid reference after deep cloning
For example if \code{TealDatasetConnector} has a dependency on some \code{TealDataset}, this
\code{TealDataset} is reassigned inside of \code{TealDatasetConnector}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TealDataAbstract$reassign_datasets_vars()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TealDataAbstract-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TealDataAbstract-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TealDataAbstract$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
