% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TealDataset.R
\name{TealDataset}
\alias{TealDataset}
\title{R6 Class representing a dataset with its attributes}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
Any \code{data.frame} object can be stored inside this object.
Some attributes like colnames, dimension or column names for a specific type will
be automatically derived.
}
\examples{

## ------------------------------------------------
## Method `TealDataset$reassign_datasets_vars`
## ------------------------------------------------

test_dataset <- teal.data:::TealDataset$new(
  dataname = "iris",
  x = iris,
  vars = list(dep = teal.data:::TealDataset$new("iris2", iris))
)
test_dataset$reassign_datasets_vars(
  list(iris2 = teal.data:::TealDataset$new("iris2", head(iris)))
)

}
\seealso{
\code{\link{MAETealDataset}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{raw_data}}{The data.frame behind this R6 class}

\item{\code{data}}{The data.frame behind this R6 class}

\item{\code{var_names}}{The column names of the data}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TealDataset-new}{\code{TealDataset$new()}}
\item \href{#method-TealDataset-recreate}{\code{TealDataset$recreate()}}
\item \href{#method-TealDataset-print}{\code{TealDataset$print()}}
\item \href{#method-TealDataset-get_dataset}{\code{TealDataset$get_dataset()}}
\item \href{#method-TealDataset-get_attrs}{\code{TealDataset$get_attrs()}}
\item \href{#method-TealDataset-get_raw_data}{\code{TealDataset$get_raw_data()}}
\item \href{#method-TealDataset-get_numeric_colnames}{\code{TealDataset$get_numeric_colnames()}}
\item \href{#method-TealDataset-get_character_colnames}{\code{TealDataset$get_character_colnames()}}
\item \href{#method-TealDataset-get_factor_colnames}{\code{TealDataset$get_factor_colnames()}}
\item \href{#method-TealDataset-get_colnames}{\code{TealDataset$get_colnames()}}
\item \href{#method-TealDataset-get_column_labels}{\code{TealDataset$get_column_labels()}}
\item \href{#method-TealDataset-get_ncol}{\code{TealDataset$get_ncol()}}
\item \href{#method-TealDataset-get_nrow}{\code{TealDataset$get_nrow()}}
\item \href{#method-TealDataset-get_rownames}{\code{TealDataset$get_rownames()}}
\item \href{#method-TealDataset-get_row_labels}{\code{TealDataset$get_row_labels()}}
\item \href{#method-TealDataset-get_dataname}{\code{TealDataset$get_dataname()}}
\item \href{#method-TealDataset-get_datanames}{\code{TealDataset$get_datanames()}}
\item \href{#method-TealDataset-get_dataset_label}{\code{TealDataset$get_dataset_label()}}
\item \href{#method-TealDataset-get_keys}{\code{TealDataset$get_keys()}}
\item \href{#method-TealDataset-get_metadata}{\code{TealDataset$get_metadata()}}
\item \href{#method-TealDataset-get_join_keys}{\code{TealDataset$get_join_keys()}}
\item \href{#method-TealDataset-get_var_r6}{\code{TealDataset$get_var_r6()}}
\item \href{#method-TealDataset-reassign_datasets_vars}{\code{TealDataset$reassign_datasets_vars()}}
\item \href{#method-TealDataset-set_dataset_label}{\code{TealDataset$set_dataset_label()}}
\item \href{#method-TealDataset-set_keys}{\code{TealDataset$set_keys()}}
\item \href{#method-TealDataset-set_join_keys}{\code{TealDataset$set_join_keys()}}
\item \href{#method-TealDataset-merge_join_keys}{\code{TealDataset$merge_join_keys()}}
\item \href{#method-TealDataset-mutate_join_keys}{\code{TealDataset$mutate_join_keys()}}
\item \href{#method-TealDataset-set_vars}{\code{TealDataset$set_vars()}}
\item \href{#method-TealDataset-set_code}{\code{TealDataset$set_code()}}
\item \href{#method-TealDataset-get_code}{\code{TealDataset$get_code()}}
\item \href{#method-TealDataset-get_code_class}{\code{TealDataset$get_code_class()}}
\item \href{#method-TealDataset-get_mutate_code_class}{\code{TealDataset$get_mutate_code_class()}}
\item \href{#method-TealDataset-get_vars}{\code{TealDataset$get_vars()}}
\item \href{#method-TealDataset-get_mutate_vars}{\code{TealDataset$get_mutate_vars()}}
\item \href{#method-TealDataset-is_mutate_delayed}{\code{TealDataset$is_mutate_delayed()}}
\item \href{#method-TealDataset-mutate}{\code{TealDataset$mutate()}}
\item \href{#method-TealDataset-check}{\code{TealDataset$check()}}
\item \href{#method-TealDataset-check_keys}{\code{TealDataset$check_keys()}}
\item \href{#method-TealDataset-is_pulled}{\code{TealDataset$is_pulled()}}
\item \href{#method-TealDataset-clone}{\code{TealDataset$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TealDataset-new"></a>}}
\if{latex}{\out{\hypertarget{method-TealDataset-new}{}}}
\subsection{Method \code{new()}}{
Create a new object of \code{TealDataset} class
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TealDataset$new(
  dataname,
  x,
  keys = character(0),
  code = character(0),
  label = character(0),
  vars = list(),
  metadata = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dataname}}{(\code{character})\cr
A given name for the dataset it may not contain spaces}

\item{\code{x}}{(\code{data.frame})\cr}

\item{\code{keys}}{optional, (\code{character})\cr
Vector with primary keys}

\item{\code{code}}{(\code{character})\cr
A character string defining the code needed to produce the data set in \code{x}.
\code{initialize()} and \code{recreate()} accept code as \code{CodeClass}
which is also needed to preserve the code uniqueness and correct order.}

\item{\code{label}}{(\code{character})\cr
Label to describe the dataset}

\item{\code{vars}}{(named \code{list})) \cr
In case when this object code depends on other \code{TealDataset} object(s) or
other constant value, this/these object(s) should be included as named
element(s) of the list. For example if this object code needs \code{ADSL}
object we should specify \verb{vars = list(ADSL = <adsl object>)}.
It is recommended to include \code{TealDataset} or \code{TealDatasetConnector} objects to
the \code{vars} list to preserve reproducibility. Please note that \code{vars}
are included to this object as local \code{vars} and they cannot be modified
within another dataset.}

\item{\code{metadata}}{(named \code{list} or \code{NULL}) \cr
Field containing metadata about the dataset. Each element of the list
should be atomic and of length one.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TealDataset-recreate"></a>}}
\if{latex}{\out{\hypertarget{method-TealDataset-recreate}{}}}
\subsection{Method \code{recreate()}}{
Recreate this \code{TealDataset} with its current attributes.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TealDataset$recreate(
  dataname = self$get_dataname(),
  x = self$get_raw_data(),
  keys = self$get_keys(),
  code = private$code,
  label = self$get_dataset_label(),
  vars = list(),
  metadata = self$get_metadata()
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dataname}}{(\code{character})\cr
A given name for the dataset it may not contain spaces}

\item{\code{x}}{(\code{data.frame})\cr}

\item{\code{keys}}{optional, (\code{character})\cr
Vector with primary keys}

\item{\code{code}}{(\code{character})\cr
A character string defining the code needed to produce the data set in \code{x}.
\code{initialize()} and \code{recreate()} accept code as \code{CodeClass}
which is also needed to preserve the code uniqueness and correct order.}

\item{\code{label}}{(\code{character})\cr
Label to describe the dataset}

\item{\code{vars}}{(named \code{list})) \cr
In case when this object code depends on other \code{TealDataset} object(s) or
other constant value, this/these object(s) should be included as named
element(s) of the list. For example if this object code needs \code{ADSL}
object we should specify \verb{vars = list(ADSL = <adsl object>)}.
It is recommended to include \code{TealDataset} or \code{TealDatasetConnector} objects to
the \code{vars} list to preserve reproducibility. Please note that \code{vars}
are included to this object as local \code{vars} and they cannot be modified
within another dataset.}

\item{\code{metadata}}{(named \code{list} or \code{NULL}) \cr
Field containing metadata about the dataset. Each element of the list
should be atomic and of length one.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a new object of the \code{TealDataset} class
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TealDataset-print"></a>}}
\if{latex}{\out{\hypertarget{method-TealDataset-print}{}}}
\subsection{Method \code{print()}}{
Prints this \code{TealDataset}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TealDataset$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{additional arguments to the printing method}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
invisibly self
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TealDataset-get_dataset"></a>}}
\if{latex}{\out{\hypertarget{method-TealDataset-get_dataset}{}}}
\subsection{Method \code{get_dataset()}}{
Performs any delayed mutate calls before returning self.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TealDataset$get_dataset()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
dataset (\code{TealDataset})
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TealDataset-get_attrs"></a>}}
\if{latex}{\out{\hypertarget{method-TealDataset-get_attrs}{}}}
\subsection{Method \code{get_attrs()}}{
Get all dataset attributes
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TealDataset$get_attrs()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
(named \code{list}) with dataset attributes
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TealDataset-get_raw_data"></a>}}
\if{latex}{\out{\hypertarget{method-TealDataset-get_raw_data}{}}}
\subsection{Method \code{get_raw_data()}}{
Derive the raw data frame inside this object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TealDataset$get_raw_data()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{data.frame}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TealDataset-get_numeric_colnames"></a>}}
\if{latex}{\out{\hypertarget{method-TealDataset-get_numeric_colnames}{}}}
\subsection{Method \code{get_numeric_colnames()}}{
Derive the names of all \code{numeric} columns
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TealDataset$get_numeric_colnames()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{character} vector.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TealDataset-get_character_colnames"></a>}}
\if{latex}{\out{\hypertarget{method-TealDataset-get_character_colnames}{}}}
\subsection{Method \code{get_character_colnames()}}{
Derive the names of all \code{character} columns
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TealDataset$get_character_colnames()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{character} vector.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TealDataset-get_factor_colnames"></a>}}
\if{latex}{\out{\hypertarget{method-TealDataset-get_factor_colnames}{}}}
\subsection{Method \code{get_factor_colnames()}}{
Derive the names of all \code{factor} columns
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TealDataset$get_factor_colnames()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{character} vector.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TealDataset-get_colnames"></a>}}
\if{latex}{\out{\hypertarget{method-TealDataset-get_colnames}{}}}
\subsection{Method \code{get_colnames()}}{
Derive the column names
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TealDataset$get_colnames()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{character} vector.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TealDataset-get_column_labels"></a>}}
\if{latex}{\out{\hypertarget{method-TealDataset-get_column_labels}{}}}
\subsection{Method \code{get_column_labels()}}{
Derive the column labels
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TealDataset$get_column_labels()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{character} vector.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TealDataset-get_ncol"></a>}}
\if{latex}{\out{\hypertarget{method-TealDataset-get_ncol}{}}}
\subsection{Method \code{get_ncol()}}{
Get the number of columns of the data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TealDataset$get_ncol()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{numeric} vector
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TealDataset-get_nrow"></a>}}
\if{latex}{\out{\hypertarget{method-TealDataset-get_nrow}{}}}
\subsection{Method \code{get_nrow()}}{
Get the number of rows of the data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TealDataset$get_nrow()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{numeric} vector
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TealDataset-get_rownames"></a>}}
\if{latex}{\out{\hypertarget{method-TealDataset-get_rownames}{}}}
\subsection{Method \code{get_rownames()}}{
Derive the row names
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TealDataset$get_rownames()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{character} vector.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TealDataset-get_row_labels"></a>}}
\if{latex}{\out{\hypertarget{method-TealDataset-get_row_labels}{}}}
\subsection{Method \code{get_row_labels()}}{
Derive the row labels
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TealDataset$get_row_labels()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{character} vector.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TealDataset-get_dataname"></a>}}
\if{latex}{\out{\hypertarget{method-TealDataset-get_dataname}{}}}
\subsection{Method \code{get_dataname()}}{
Derive the \code{name} which was formerly called \code{dataname}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TealDataset$get_dataname()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{character} name of the dataset
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TealDataset-get_datanames"></a>}}
\if{latex}{\out{\hypertarget{method-TealDataset-get_datanames}{}}}
\subsection{Method \code{get_datanames()}}{
Derive the dataname
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TealDataset$get_datanames()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{character} name of the dataset
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TealDataset-get_dataset_label"></a>}}
\if{latex}{\out{\hypertarget{method-TealDataset-get_dataset_label}{}}}
\subsection{Method \code{get_dataset_label()}}{
Derive the \code{label} which was former called \code{datalabel}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TealDataset$get_dataset_label()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{character} label of the dataset
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TealDataset-get_keys"></a>}}
\if{latex}{\out{\hypertarget{method-TealDataset-get_keys}{}}}
\subsection{Method \code{get_keys()}}{
Get primary keys of dataset
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TealDataset$get_keys()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
(\code{character} vector) with dataset primary keys
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TealDataset-get_metadata"></a>}}
\if{latex}{\out{\hypertarget{method-TealDataset-get_metadata}{}}}
\subsection{Method \code{get_metadata()}}{
Get metadata of dataset
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TealDataset$get_metadata()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
(named \code{list})
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TealDataset-get_join_keys"></a>}}
\if{latex}{\out{\hypertarget{method-TealDataset-get_join_keys}{}}}
\subsection{Method \code{get_join_keys()}}{
Get \code{JoinKeys} object with keys used for joining.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TealDataset$get_join_keys()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
(\code{JoinKeys})
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TealDataset-get_var_r6"></a>}}
\if{latex}{\out{\hypertarget{method-TealDataset-get_var_r6}{}}}
\subsection{Method \code{get_var_r6()}}{
Get the list of dependencies that are \code{TealDataset} or \code{TealDatasetConnector} objects
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TealDataset$get_var_r6()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{list}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TealDataset-reassign_datasets_vars"></a>}}
\if{latex}{\out{\hypertarget{method-TealDataset-reassign_datasets_vars}{}}}
\subsection{Method \code{reassign_datasets_vars()}}{
Overwrites \code{TealDataset} or \code{TealDatasetConnector} dependencies of this \code{TealDataset} with
those found in \code{datasets}. Reassignment
refers only to the provided \code{datasets}, other \code{vars} remains the same.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TealDataset$reassign_datasets_vars(datasets)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{datasets}}{(\verb{named list} of \code{TealDataset(s)} or \code{TealDatasetConnector(s)})\cr
objects with valid pointers.}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
Reassign \code{vars} in this object to keep references up to date after deep clone.
Update is done based on the objects passed in \code{datasets} argument.
Overwrites dependencies with names matching the names of the objects passed
in \code{datasets}.
}

\subsection{Returns}{
NULL invisible
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{test_dataset <- teal.data:::TealDataset$new(
  dataname = "iris",
  x = iris,
  vars = list(dep = teal.data:::TealDataset$new("iris2", iris))
)
test_dataset$reassign_datasets_vars(
  list(iris2 = teal.data:::TealDataset$new("iris2", head(iris)))
)

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TealDataset-set_dataset_label"></a>}}
\if{latex}{\out{\hypertarget{method-TealDataset-set_dataset_label}{}}}
\subsection{Method \code{set_dataset_label()}}{
Set the label for the dataset
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TealDataset$set_dataset_label(label)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{label}}{(\code{character})\cr
Label to describe the dataset}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
(\code{self}) invisibly for chaining
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TealDataset-set_keys"></a>}}
\if{latex}{\out{\hypertarget{method-TealDataset-set_keys}{}}}
\subsection{Method \code{set_keys()}}{
Set new keys
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TealDataset$set_keys(keys)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{keys}}{optional, (\code{character})\cr
Vector with primary keys}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
(\code{self}) invisibly for chaining.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TealDataset-set_join_keys"></a>}}
\if{latex}{\out{\hypertarget{method-TealDataset-set_join_keys}{}}}
\subsection{Method \code{set_join_keys()}}{
set join_keys for a given dataset and object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TealDataset$set_join_keys(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{\code{list} of \code{JoinKeySet} objects (which are created using the \code{join_key} function)
or single \code{JoinKeySet} objects}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
(\code{self}) invisibly for chaining
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TealDataset-merge_join_keys"></a>}}
\if{latex}{\out{\hypertarget{method-TealDataset-merge_join_keys}{}}}
\subsection{Method \code{merge_join_keys()}}{
merge input join key with join key inside of object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TealDataset$merge_join_keys(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{\code{list} of \code{JoinKeys} objects or single \code{JoinKeys} object}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
(\code{self}) invisibly for chaining
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TealDataset-mutate_join_keys"></a>}}
\if{latex}{\out{\hypertarget{method-TealDataset-mutate_join_keys}{}}}
\subsection{Method \code{mutate_join_keys()}}{
mutate the join_keys for a given dataset and object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TealDataset$mutate_join_keys(dataset, val)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dataset}}{(\code{character}) dataset for which join_keys are to be set against self}

\item{\code{val}}{(named \code{character}) column names used to join}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
(\code{self}) invisibly for chaining
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TealDataset-set_vars"></a>}}
\if{latex}{\out{\hypertarget{method-TealDataset-set_vars}{}}}
\subsection{Method \code{set_vars()}}{
Adds variables which code depends on
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TealDataset$set_vars(vars)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{vars}}{(\verb{named list}) contains any R object which code depends on}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
(\code{self}) invisibly for chaining
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TealDataset-set_code"></a>}}
\if{latex}{\out{\hypertarget{method-TealDataset-set_code}{}}}
\subsection{Method \code{set_code()}}{
Sets reproducible code
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TealDataset$set_code(code)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{code}}{(\code{character})\cr
A character string defining the code needed to produce the data set in \code{x}.
\code{initialize()} and \code{recreate()} accept code as \code{CodeClass}
which is also needed to preserve the code uniqueness and correct order.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
(\code{self}) invisibly for chaining
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TealDataset-get_code"></a>}}
\if{latex}{\out{\hypertarget{method-TealDataset-get_code}{}}}
\subsection{Method \code{get_code()}}{
Get code to get data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TealDataset$get_code(deparse = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deparse}}{(\code{logical}) whether return deparsed form of a call}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
optionally deparsed \code{call} object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TealDataset-get_code_class"></a>}}
\if{latex}{\out{\hypertarget{method-TealDataset-get_code_class}{}}}
\subsection{Method \code{get_code_class()}}{
Get internal \code{CodeClass} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TealDataset$get_code_class(nodeps = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{nodeps}}{(\code{logical(1)}) whether \code{CodeClass} should not contain the code
of the dependent \code{vars}
the \code{mutate}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{CodeClass}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TealDataset-get_mutate_code_class"></a>}}
\if{latex}{\out{\hypertarget{method-TealDataset-get_mutate_code_class}{}}}
\subsection{Method \code{get_mutate_code_class()}}{
Get internal \code{CodeClass} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TealDataset$get_mutate_code_class()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{CodeClass}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TealDataset-get_vars"></a>}}
\if{latex}{\out{\hypertarget{method-TealDataset-get_vars}{}}}
\subsection{Method \code{get_vars()}}{
Get internal \code{vars} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TealDataset$get_vars()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{list}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TealDataset-get_mutate_vars"></a>}}
\if{latex}{\out{\hypertarget{method-TealDataset-get_mutate_vars}{}}}
\subsection{Method \code{get_mutate_vars()}}{
Get internal \code{mutate_vars} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TealDataset$get_mutate_vars()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{list}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TealDataset-is_mutate_delayed"></a>}}
\if{latex}{\out{\hypertarget{method-TealDataset-is_mutate_delayed}{}}}
\subsection{Method \code{is_mutate_delayed()}}{
Whether mutate code has delayed evaluation.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TealDataset$is_mutate_delayed()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{logical}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TealDataset-mutate"></a>}}
\if{latex}{\out{\hypertarget{method-TealDataset-mutate}{}}}
\subsection{Method \code{mutate()}}{
Mutate dataset by code
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TealDataset$mutate(code, vars = list(), force_delay = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{code}}{(\code{CodeClass}) or (\code{character}) R expressions to be executed}

\item{\code{vars}}{a named list of R objects that \code{code} depends on to execute}

\item{\code{force_delay}}{(\code{logical}) used by the containing \code{TealDatasetConnector} object

Either code or script must be provided, but not both.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
(\code{self}) invisibly for chaining
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TealDataset-check"></a>}}
\if{latex}{\out{\hypertarget{method-TealDataset-check}{}}}
\subsection{Method \code{check()}}{
Check to determine if the raw data is reproducible from the \code{get_code()} code.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TealDataset$check()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{TRUE} if the dataset generated from evaluating the
\code{get_code()} code is identical to the raw data, else \code{FALSE}.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TealDataset-check_keys"></a>}}
\if{latex}{\out{\hypertarget{method-TealDataset-check_keys}{}}}
\subsection{Method \code{check_keys()}}{
Check if keys has been specified correctly for dataset. Set of \code{keys}
should distinguish unique rows or be \code{character(0)}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TealDataset$check_keys(keys = private$.keys)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{keys}}{optional, (\code{character})\cr
Vector with primary keys}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
\code{TRUE} if dataset has been already pulled, else \code{FALSE}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TealDataset-is_pulled"></a>}}
\if{latex}{\out{\hypertarget{method-TealDataset-is_pulled}{}}}
\subsection{Method \code{is_pulled()}}{
Check if dataset has already been pulled.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TealDataset$is_pulled()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
\code{TRUE} if dataset has been already pulled, else \code{FALSE}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TealDataset-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TealDataset-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TealDataset$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
