% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testhat-helpers.R
\name{test_join_keys_add}
\alias{test_join_keys_add}
\title{Test suite for join_keys after manual adding a primary key}
\usage{
test_join_keys_add(
  obj,
  dataset_1 = "ds1",
  new_dataset_1 = "ds2",
  new_keys = c("id")
)
}
\arguments{
\item{obj}{Object with \code{join_keys} to perform operations (can be \code{join_keys}
or \code{teal_data})}

\item{dataset_1}{\code{character(1)} name of existing dataset.}

\item{new_dataset_1}{\code{character(1)} name of new dataset to add.}

\item{new_keys}{}
}
\value{
\code{obj} itself modified with a new foreign key.
}
\description{
Test suite for join_keys after manual adding a primary key
}
