% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testhat-helpers.R
\name{test_join_keys_bare}
\alias{test_join_keys_bare}
\title{Test suite for default join_keys generated by helper}
\usage{
test_join_keys_bare(obj, dataset_1 = "ds1")
}
\arguments{
\item{obj}{Object with \code{join_keys} to perform operations (can be \code{join_keys}
or \code{teal_data}). It should only contain a primary key.}

\item{dataset_1}{\code{character(1)} name of existing dataset to test.}
}
\value{
\code{obj} itself without any modifications
}
\description{
Test suite for default join_keys generated by helper
}
