% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testhat-helpers.R
\name{test_join_keys_combinatory}
\alias{test_join_keys_combinatory}
\title{(test helper) Test suite for join_keys that performs a mass modification}
\usage{
test_join_keys_combinatory(obj)
}
\arguments{
\item{obj}{Object with \code{join_keys} to perform operations (can be \code{join_keys}
or \code{teal_data})}
}
\value{
\code{obj} itself modified with a new foreign key.
}
\description{
The goal of this helper is to modify the \code{join_keys} with all variants of a
valid foreign key.
}
\keyword{internal}
