% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_relational_data.R
\name{to_relational_data}
\alias{to_relational_data}
\title{S3 generic for \code{to_relational_data} function.}
\usage{
to_relational_data(data)
}
\arguments{
\item{data}{\code{TealDataset}, \code{TealDatasetConnector}, \code{data.frame}, \code{MultiAssayExperiment},  \code{list}
or \code{function} returning a named list.}
}
\value{
\code{TealData} object
}
\description{
This function takes an object and converts into a \code{TealData} object, the primary data
object for use in teal applications.
}
\details{
Passing a \code{TealData} into this function leaves the object unchanged.
}
\examples{

to_relational_data(head(iris))
to_relational_data(dataset("IRIS", head(iris)))
to_relational_data(list(iris = head(iris), mtcars = head(mtcars)))

d_connector <- dataset_connector("iris", callable_function(function() head(iris)))
d_connector$pull()
to_relational_data(d_connector)

}
\keyword{internal}
