% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teal_data.R
\name{update_join_keys_to_primary}
\alias{update_join_keys_to_primary}
\title{Add primary keys as join_keys to a dataset self}
\usage{
update_join_keys_to_primary(data_objects, x)
}
\arguments{
\item{data_objects}{(\code{list}) of \code{TealDataset}, \code{TealDatasetConnector} or \code{TealDataConnector} objects}

\item{x}{(\code{join_keys}) object}
}
\description{
Add primary keys as join_keys to a dataset self
}
\keyword{internal}
