% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddl-username_password_module.R
\name{username_password_module}
\alias{username_password_module}
\alias{username_password_server}
\alias{username_password_ui}
\alias{username_password_args}
\title{offline_args for username_password_module}
\usage{
username_password_server(id, offline_args, code, postprocess_fun)

username_password_ui(id)

username_password_args()
}
\arguments{
\item{id}{(\code{character}) \code{shiny} module id.}

\item{offline_args}{(\code{list} named)\cr
arguments to be substituted in the \code{code}. These
argument are going to replace arguments set through
\code{ui} and \code{server}. Example use case is when app user
is asked to input a password and we'd like to skip this
input in the reproducible code. Typically users password
is substituted with \code{askpass::askpass()} call, so the
returned code is still executable but secure.}

\item{code}{(\code{character})\cr
Code to be evaluated and returned to the \code{postprocess_fun}}

\item{postprocess_fun}{(\verb{function(env, code)})\cr
Function to be run after code is run. This function suppose
has two arguments:
\itemize{
\item \code{env} (\code{environment}) returned as a result of the code evaluation
\item code (\code{character}) \code{code} provided with resolved (substituted) args.
}}
}
\description{
offline_args for username_password_module
}
