% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_metadata}
\alias{validate_metadata}
\title{validate metadata as a list of length one atomic entries (or NULL)}
\usage{
validate_metadata(metadata)
}
\arguments{
\item{metadata}{\code{object} to be checked}
}
\value{
\code{NULL} or throw error
}
\description{
validate metadata as a list of length one atomic entries (or NULL)
}
\examples{

validate_metadata(NULL)
validate_metadata(list(A = TRUE, B = 10, C = "test"))
\dontrun{
validate_metadata(list(a = 1:10))
}

}
