% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_keys-utils.R
\name{assert_compatible_keys2}
\alias{assert_compatible_keys2}
\title{Helper function to ensuring compatibility between two sets of keys}
\usage{
assert_compatible_keys2(x, y)
}
\value{
Returns TRUE if successful, otherwise raises error.
}
\description{
Helper function to ensuring compatibility between two sets of keys
}
\keyword{internal}
