% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdisc_data.R
\name{cdisc_data}
\alias{cdisc_data}
\title{Data input for teal app}
\usage{
cdisc_data(..., join_keys = teal.data::join_keys(), code = "", check = FALSE)
}
\arguments{
\item{...}{(\code{TealDataConnector}, \code{TealDataset} or
\code{TealDatasetConnector}) elements to include where \code{ADSL} data is mandatory.}

\item{join_keys}{(\code{JoinKeys}) or a single (\code{JoinKeySet})\cr
(optional) object with datasets column names used for joining.
If empty then it would be automatically derived basing on intersection of datasets primary keys}

\item{code}{(\code{character}) code to reproduce the datasets.}

\item{check}{(\code{logical}) reproducibility check - whether to perform a check that the pre-processing
code included in the object definitions actually produces those objects.
If \code{check} is true and preprocessing code is empty an error will be thrown.}
}
\value{
a \code{TealData} object
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
Function takes datasets and creates \code{TealData} object which can be used in \code{teal} applications.
}
\details{
This function checks if there were keys added to all data sets
}
\note{
This function does not automatically assign keys to \code{TealDataset}
and \code{TealDatasetConnector} objects passed to it. If the keys are needed
they should be assigned before calling \code{cdisc_data}. See example:

\if{html}{\out{<div class="sourceCode">}}\preformatted{library(scda)
test_dataset <- dataset("ADAE", synthetic_cdisc_data("latest")$adae) # does not have keys
test_adsl <- cdisc_dataset("ADSL", synthetic_cdisc_data("latest")$adsl)
test_data <- cdisc_data(test_dataset, test_adsl)
get_keys(test_data, "ADAE") # returns character(0)

test_dataset <- cdisc_dataset("ADAE", synthetic_cdisc_data("latest")$adae)
test_data <- cdisc_data(test_dataset, test_adsl)
get_keys(test_data, "ADAE") # returns [1] "STUDYID" "USUBJID" "ASTDTM"  "AETERM"  "AESEQ"
}\if{html}{\out{</div>}}
}
\examples{
library(scda)

latest_data <- synthetic_cdisc_data("latest")
ADSL <- latest_data$adsl
ADTTE <- latest_data$adtte

# basic example
cdisc_data(
  cdisc_dataset("ADSL", ADSL),
  cdisc_dataset("ADTTE", ADTTE),
  code = "ADSL <- synthetic_cdisc_data('latest')$adsl
          ADTTE <- synthetic_cdisc_data('latest')$adtte",
  check = TRUE
)

# Example with keys
cdisc_data(
  cdisc_dataset("ADSL", ADSL,
    keys = c("STUDYID", "USUBJID")
  ),
  cdisc_dataset("ADTTE", ADTTE,
    keys = c("STUDYID", "USUBJID", "PARAMCD"),
    parent = "ADSL"
  ),
  join_keys = join_keys(
    join_key(
      "ADSL",
      "ADTTE",
      c("STUDYID" = "STUDYID", "USUBJID" = "USUBJID")
    )
  ),
  code = "ADSL <- synthetic_cdisc_data('latest')$adsl
          ADTTE <- synthetic_cdisc_data('latest')$adtte",
  check = TRUE
)
}
