% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdisc_data.R
\name{cdisc_data}
\alias{cdisc_data}
\title{Data input for teal app}
\usage{
cdisc_data(
  ...,
  join_keys = teal.data::cdisc_join_keys(...),
  code = character(0),
  check = FALSE
)
}
\arguments{
\item{...}{(\code{TealDataConnector}, \code{TealDataset}, \code{TealDatasetConnector}, \code{any})\cr
Either 1) an object of a \verb{Teal*} class, which is deprecated and will be removed in next release,
or 2) any number of any objects provided as \code{name = value} pairs, which is available from version \verb{0.4.0}.}

\item{join_keys}{(\code{JoinKeys}) or a single (\code{join_key_set})\cr
(optional) object with datasets column names used for joining.
If empty then it would be automatically derived basing on intersection of datasets primary keys.
For ADAM datasets it would be automatically derived.}

\item{code}{(\code{character}, \code{language}) code to reproduce the datasets.}

\item{check}{(\code{logical}) reproducibility check - whether to perform a check that the pre-processing
code included in the object definitions actually produces those objects.
If \code{check} is true and preprocessing code is empty an error will be thrown.}
}
\value{
\itemize{
\item a \code{TealData} object when \code{TealDataset} \code{TealDatasetConnector}, \code{TealDataConnector} is provided,
\item a \code{teal_data} object otherwise.
}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
Function is a wrapper around \code{\link[=teal_data]{teal_data()}} and guesses \code{join_keys}
for given datasets whose names match ADAM datasets names.
}
\details{
This function checks if there were keys added to all data sets
}
\examples{
data <- cdisc_data(
  join_keys = join_keys(
    join_key("ADSL", "ADTTE", c("STUDYID" = "STUDYID", "USUBJID" = "USUBJID"))
  )
)

data <- within(data, {
  ADSL <- teal.data::example_cdisc_data("ADSL")
  ADTTE <- teal.data::example_cdisc_data("ADTTE")
})

}
