% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CDISCTealDataConnector.R
\name{CDISCTealDataConnector}
\alias{CDISCTealDataConnector}
\title{Manage multiple and \code{TealDatasetConnector} of the same type.}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
Class manages \code{TealDatasetConnector} to specify additional dynamic arguments and to
open/close connection.
}
\section{Super classes}{
\code{\link[teal.data:TealDataAbstract]{teal.data::TealDataAbstract}} -> \code{\link[teal.data:TealDataConnector]{teal.data::TealDataConnector}} -> \code{CDISCTealDataConnector}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-CDISCTealDataConnector-new}{\code{CDISCTealDataConnector$new()}}
\item \href{#method-CDISCTealDataConnector-get_parent}{\code{CDISCTealDataConnector$get_parent()}}
\item \href{#method-CDISCTealDataConnector-clone}{\code{CDISCTealDataConnector$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="teal.data" data-topic="TealDataAbstract" data-id="check"><a href='../../teal.data/html/TealDataAbstract.html#method-TealDataAbstract-check'><code>teal.data::TealDataAbstract$check()</code></a></span></li>
<li><span class="pkg-link" data-pkg="teal.data" data-topic="TealDataAbstract" data-id="check_reproducibility"><a href='../../teal.data/html/TealDataAbstract.html#method-TealDataAbstract-check_reproducibility'><code>teal.data::TealDataAbstract$check_reproducibility()</code></a></span></li>
<li><span class="pkg-link" data-pkg="teal.data" data-topic="TealDataAbstract" data-id="execute_mutate"><a href='../../teal.data/html/TealDataAbstract.html#method-TealDataAbstract-execute_mutate'><code>teal.data::TealDataAbstract$execute_mutate()</code></a></span></li>
<li><span class="pkg-link" data-pkg="teal.data" data-topic="TealDataAbstract" data-id="get_check"><a href='../../teal.data/html/TealDataAbstract.html#method-TealDataAbstract-get_check'><code>teal.data::TealDataAbstract$get_check()</code></a></span></li>
<li><span class="pkg-link" data-pkg="teal.data" data-topic="TealDataAbstract" data-id="get_check_result"><a href='../../teal.data/html/TealDataAbstract.html#method-TealDataAbstract-get_check_result'><code>teal.data::TealDataAbstract$get_check_result()</code></a></span></li>
<li><span class="pkg-link" data-pkg="teal.data" data-topic="TealDataAbstract" data-id="get_code"><a href='../../teal.data/html/TealDataAbstract.html#method-TealDataAbstract-get_code'><code>teal.data::TealDataAbstract$get_code()</code></a></span></li>
<li><span class="pkg-link" data-pkg="teal.data" data-topic="TealDataAbstract" data-id="get_datanames"><a href='../../teal.data/html/TealDataAbstract.html#method-TealDataAbstract-get_datanames'><code>teal.data::TealDataAbstract$get_datanames()</code></a></span></li>
<li><span class="pkg-link" data-pkg="teal.data" data-topic="TealDataAbstract" data-id="get_dataset"><a href='../../teal.data/html/TealDataAbstract.html#method-TealDataAbstract-get_dataset'><code>teal.data::TealDataAbstract$get_dataset()</code></a></span></li>
<li><span class="pkg-link" data-pkg="teal.data" data-topic="TealDataAbstract" data-id="get_datasets"><a href='../../teal.data/html/TealDataAbstract.html#method-TealDataAbstract-get_datasets'><code>teal.data::TealDataAbstract$get_datasets()</code></a></span></li>
<li><span class="pkg-link" data-pkg="teal.data" data-topic="TealDataAbstract" data-id="get_items"><a href='../../teal.data/html/TealDataAbstract.html#method-TealDataAbstract-get_items'><code>teal.data::TealDataAbstract$get_items()</code></a></span></li>
<li><span class="pkg-link" data-pkg="teal.data" data-topic="TealDataAbstract" data-id="is_pulled"><a href='../../teal.data/html/TealDataAbstract.html#method-TealDataAbstract-is_pulled'><code>teal.data::TealDataAbstract$is_pulled()</code></a></span></li>
<li><span class="pkg-link" data-pkg="teal.data" data-topic="TealDataAbstract" data-id="mutate_dataset"><a href='../../teal.data/html/TealDataAbstract.html#method-TealDataAbstract-mutate_dataset'><code>teal.data::TealDataAbstract$mutate_dataset()</code></a></span></li>
<li><span class="pkg-link" data-pkg="teal.data" data-topic="TealDataAbstract" data-id="reassign_datasets_vars"><a href='../../teal.data/html/TealDataAbstract.html#method-TealDataAbstract-reassign_datasets_vars'><code>teal.data::TealDataAbstract$reassign_datasets_vars()</code></a></span></li>
<li><span class="pkg-link" data-pkg="teal.data" data-topic="TealDataAbstract" data-id="set_check"><a href='../../teal.data/html/TealDataAbstract.html#method-TealDataAbstract-set_check'><code>teal.data::TealDataAbstract$set_check()</code></a></span></li>
<li><span class="pkg-link" data-pkg="teal.data" data-topic="TealDataAbstract" data-id="set_pull_code"><a href='../../teal.data/html/TealDataAbstract.html#method-TealDataAbstract-set_pull_code'><code>teal.data::TealDataAbstract$set_pull_code()</code></a></span></li>
<li><span class="pkg-link" data-pkg="teal.data" data-topic="TealDataConnector" data-id="get_code_class"><a href='../../teal.data/html/TealDataConnector.html#method-TealDataConnector-get_code_class'><code>teal.data::TealDataConnector$get_code_class()</code></a></span></li>
<li><span class="pkg-link" data-pkg="teal.data" data-topic="TealDataConnector" data-id="get_connection"><a href='../../teal.data/html/TealDataConnector.html#method-TealDataConnector-get_connection'><code>teal.data::TealDataConnector$get_connection()</code></a></span></li>
<li><span class="pkg-link" data-pkg="teal.data" data-topic="TealDataConnector" data-id="get_preopen_server"><a href='../../teal.data/html/TealDataConnector.html#method-TealDataConnector-get_preopen_server'><code>teal.data::TealDataConnector$get_preopen_server()</code></a></span></li>
<li><span class="pkg-link" data-pkg="teal.data" data-topic="TealDataConnector" data-id="get_server"><a href='../../teal.data/html/TealDataConnector.html#method-TealDataConnector-get_server'><code>teal.data::TealDataConnector$get_server()</code></a></span></li>
<li><span class="pkg-link" data-pkg="teal.data" data-topic="TealDataConnector" data-id="get_ui"><a href='../../teal.data/html/TealDataConnector.html#method-TealDataConnector-get_ui'><code>teal.data::TealDataConnector$get_ui()</code></a></span></li>
<li><span class="pkg-link" data-pkg="teal.data" data-topic="TealDataConnector" data-id="is_failed"><a href='../../teal.data/html/TealDataConnector.html#method-TealDataConnector-is_failed'><code>teal.data::TealDataConnector$is_failed()</code></a></span></li>
<li><span class="pkg-link" data-pkg="teal.data" data-topic="TealDataConnector" data-id="launch"><a href='../../teal.data/html/TealDataConnector.html#method-TealDataConnector-launch'><code>teal.data::TealDataConnector$launch()</code></a></span></li>
<li><span class="pkg-link" data-pkg="teal.data" data-topic="TealDataConnector" data-id="mutate"><a href='../../teal.data/html/TealDataConnector.html#method-TealDataConnector-mutate'><code>teal.data::TealDataConnector$mutate()</code></a></span></li>
<li><span class="pkg-link" data-pkg="teal.data" data-topic="TealDataConnector" data-id="print"><a href='../../teal.data/html/TealDataConnector.html#method-TealDataConnector-print'><code>teal.data::TealDataConnector$print()</code></a></span></li>
<li><span class="pkg-link" data-pkg="teal.data" data-topic="TealDataConnector" data-id="pull"><a href='../../teal.data/html/TealDataConnector.html#method-TealDataConnector-pull'><code>teal.data::TealDataConnector$pull()</code></a></span></li>
<li><span class="pkg-link" data-pkg="teal.data" data-topic="TealDataConnector" data-id="set_preopen_server"><a href='../../teal.data/html/TealDataConnector.html#method-TealDataConnector-set_preopen_server'><code>teal.data::TealDataConnector$set_preopen_server()</code></a></span></li>
<li><span class="pkg-link" data-pkg="teal.data" data-topic="TealDataConnector" data-id="set_pull_args"><a href='../../teal.data/html/TealDataConnector.html#method-TealDataConnector-set_pull_args'><code>teal.data::TealDataConnector$set_pull_args()</code></a></span></li>
<li><span class="pkg-link" data-pkg="teal.data" data-topic="TealDataConnector" data-id="set_server"><a href='../../teal.data/html/TealDataConnector.html#method-TealDataConnector-set_server'><code>teal.data::TealDataConnector$set_server()</code></a></span></li>
<li><span class="pkg-link" data-pkg="teal.data" data-topic="TealDataConnector" data-id="set_ui"><a href='../../teal.data/html/TealDataConnector.html#method-TealDataConnector-set_ui'><code>teal.data::TealDataConnector$set_ui()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CDISCTealDataConnector-new"></a>}}
\if{latex}{\out{\hypertarget{method-CDISCTealDataConnector-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{CDISCTealDataConnector} object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CDISCTealDataConnector$new(connection, connectors)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{connection}}{(\code{TealDataConnection})\cr
connection to data source}

\item{\code{connectors}}{(\code{list} of \code{TealDatasetConnector} elements)\cr
list with dataset connectors}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CDISCTealDataConnector-get_parent"></a>}}
\if{latex}{\out{\hypertarget{method-CDISCTealDataConnector-get_parent}{}}}
\subsection{Method \code{get_parent()}}{
Get all datasets parent names
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CDISCTealDataConnector$get_parent()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
(named \code{list}) with dataset name and its corresponding parent dataset name
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CDISCTealDataConnector-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CDISCTealDataConnector-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CDISCTealDataConnector$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
