% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddl.R
\name{ddl}
\alias{ddl}
\title{DDL object}
\usage{
ddl(
  code,
  ui = submit_button_ui,
  server = submit_button_server,
  offline_args = list(),
  postprocess_fun = function(env_list, code, join_keys) {
    
    do.call(teal.data::teal_data, args = c(env_list, code = code, join_keys =
    join_keys))
 },
  join_keys = teal.data::join_keys(),
  datanames
)
}
\arguments{
\item{code}{(\code{character})\cr
Code to be evaluated and returned to the \code{postprocess_fun}}

\item{ui}{(\code{shiny.tag})\cr
\code{shiny} ui module containing inputs which \code{id} correspond to the
args in the \code{code}.}

\item{server}{(\verb{function(id, offline_args, code, postprocess_fun)})\cr
\code{shiny} server module returning data. This server suppose to execute
DDL code and return a reactive data containing necessary data.
Package provides universal \code{username_password_server} which
runs \link{ddl_run} function, which returns \code{tdata} object.
Details in the the example}

\item{offline_args}{(\code{list} named)\cr
arguments to be substituted in the \code{code}. These
argument are going to replace arguments set through
\code{ui} and \code{server}. Example use case is when app user
is asked to input a password and we'd like to skip this
input in the reproducible code. Typically users password
is substituted with \code{askpass::askpass()} call, so the
returned code is still executable but secure.}

\item{postprocess_fun}{(\verb{function(env, code)})\cr
Function to be run after code is run. This function suppose
has two arguments:
\itemize{
\item \code{env} (\code{environment}) returned as a result of the code evaluation
\item code (\code{character}) \code{code} provided with resolved (substituted) args.
}}

\item{join_keys}{(\code{JoinKeys}) or a single (\code{JoinKeySet})\cr
(optional) object with dataset column relationships used for joining.
If empty then no joins between pairs of objects}

\item{datanames}{(\code{character})\cr
Names of the objects to be created from the code evaluation.
If not specified (\code{character(0)}), all objects will be used to \code{teal_data} function
(via \code{env_list} in \code{postprocess_fun}).}
}
\description{
Object to execute custom DDL code in the shiny session
}
