% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddl.R
\name{ddl_run}
\alias{ddl_run}
\title{Creates \code{tdata} object}
\usage{
ddl_run(offline_args, code, postprocess_fun, input)
}
\arguments{
\item{offline_args}{(\code{list} named)\cr
arguments to be substituted in the \code{code}. These
argument are going to replace arguments set through
\code{ui} and \code{server}. Example use case is when app user
is asked to input a password and we'd like to skip this
input in the reproducible code. Typically users password
is substituted with \code{askpass::askpass()} call, so the
returned code is still executable but secure.}

\item{code}{(\code{character})\cr
Code to be evaluated and returned to the \code{postprocess_fun}}

\item{postprocess_fun}{(\verb{function(env, code)})\cr
Function to be run after code is run. This function suppose
has two arguments:
\itemize{
\item \code{env} (\code{environment}) returned as a result of the code evaluation
\item code (\code{character}) \code{code} provided with resolved (substituted) args.
}}
}
\description{
Resolves arguments and executes custom DDL \code{code}.
Custom \code{code} and \code{data} created from code evaluation
are passed to the \code{postprocess_fun}
}
