% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_cdisc_join_keys.R
\docType{data}
\name{default_cdisc_join_keys}
\alias{default_cdisc_join_keys}
\title{List containing the default \code{CDISC} join keys}
\description{
List containing the default \code{CDISC} join keys
}
\details{
This data object is created at loading time from \code{cdisc_datasets/cdisc_datasets.yaml}.
}
