% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_cdisc_join_keys.R
\docType{data}
\name{default_cdisc_join_keys}
\alias{default_cdisc_join_keys}
\title{List containing default joining keys for \code{CDISC} datasets}
\description{
This data object is created at loading time from \code{cdisc_datasets/cdisc_datasets.yaml}.
}
