% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{default_cdisc_join_keys}
\alias{default_cdisc_join_keys}
\title{List containing default joining keys for \code{CDISC} datasets}
\format{
An object of class \code{join_keys} (inherits from \code{list}) of length 19.
}
\source{
internal. See \code{data-raw/cdisc_datasets.yaml} and \code{data-raw/data.R} for details.
}
\usage{
default_cdisc_join_keys
}
\description{
These represent primary and foreign keys for joining \code{CDISC} datasets.
}
\examples{
default_cdisc_join_keys
default_cdisc_join_keys[c("ADSL", "ADAE")] # Subset for ADSL and ADAE
}
\seealso{
\code{\link[=join_keys]{join_keys()}}
}
\keyword{datasets}
