% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdisc_data.R
\name{deprecated_join_keys_extract}
\alias{deprecated_join_keys_extract}
\title{Extrapolate parents from \code{TealData} classes}
\usage{
deprecated_join_keys_extract(data_objects, join_keys)
}
\description{
note: This function will be removed once the following classes are defunct:
\code{TealDataConnector}, \code{TealDataset}, \code{TealDatasetConnector}
}
\keyword{internal}
