% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_keys-utils.R
\name{.append_indirect_links}
\alias{.append_indirect_links}
\title{Append indirect links}
\usage{
.append_indirect_links(x)
}
\arguments{
\item{x}{(\code{join_keys}) object to update the keys.}
}
\value{
(\code{self}) invisibly for chaining
}
\description{
Adds the keys between datasets if they are connected by the same set of keys
via other foreign datasets.
}
\keyword{internal}
