% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tdata_ddl-dummy_teal_init.R
\name{dummy_teal_init}
\alias{dummy_teal_init}
\title{Dummy DDL app}
\usage{
dummy_teal_init(object)
}
\description{
Run it
}
\examples{

example1 <- tdata_ddl(
  # code to be run when app user presses submit
  code = "
    conn <- open_dummy_connection(username = {username}, password = {password})
    my_data <- get_dummy_data(conn)
    close_conn()
  ",

  # arguments used for show R code
  offline_args = username_password_args(),

  # ui they wish to use for the loading data
  ui = username_password_ui,


  tdata_function = function(env_list, code) {
    new_tdata(data = list(DATA = env_list$my_data), code = code)
  }
)
app <- dummy_teal_init(example1)

if (interactive()) {
  shiny::runApp(app)
}
}
