% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy_function.R
\name{example_cdisc_data}
\alias{example_cdisc_data}
\title{Generate sample \code{CDISC} datasets for testing \code{teal} application}
\usage{
example_cdisc_data(dataname)
}
\arguments{
\item{dataname}{name of the \code{CDISC} dataset}
}
\value{
\code{cdisc_data}
}
\description{
This function returns a dummy dataset for testing examples and should only be used within \code{teal.data}.
It is not meant to retrieve the \code{SCDA} dataset, and the dataset itself is not maintained here.
}
\details{
This function creates a copy of the \code{SCDA} data for testing purposes.

\code{CDISC} data includes \code{ADSL}, \code{ADAE}, \code{ADLB}, \code{ADCM}, \code{ADEX}, \code{ADRS}, \code{ADTR} and \code{ADTTE}.
}
