% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teal_data-get_code.R
\name{get_code,teal_data-method}
\alias{get_code,teal_data-method}
\title{Get Code from \code{teal_data}}
\usage{
\S4method{get_code}{teal_data}(object, deparse = TRUE, names = NULL)
}
\arguments{
\item{object}{\code{teal_data}}

\item{deparse}{(\code{logical}) whether return deparsed form of a call.}

\item{names}{(\code{character}) vector of object names to return the code for.}
}
\value{
(\code{character}) code of import and preparation of data for \code{teal} application.
}
\description{
Get code from \code{\link{teal_data}}. The extraction of the code limited to specific objects can be performed with the usage
of \code{names} parameter. If input \code{teal_data@verified} is \code{FALSE} the returned code is pre-pended with a warning
stating that the object was not verified.
}
\examples{

tdata1 <- teal_data()
tdata1 <- within(tdata1, {
  a <- 1
  b <- a^5
  c <- list(x = 2)
})
get_code(tdata1)
datanames(tdata1) <- c("a", "b", "c")
get_code(tdata1, names = "a")
get_code(tdata1, names = "b")

tdata2 <- teal_data(x1 = iris, code = "x1 = iris")
get_code(tdata2)
get_code(verify(tdata2))
}
