% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teal_data-get_code.R
\name{get_code,teal_data-method}
\alias{get_code,teal_data-method}
\title{Get Code from \code{teal_data}}
\usage{
\S4method{get_code}{teal_data}(object, deparse = TRUE, datanames = NULL)
}
\arguments{
\item{object}{(\code{teal_data})}

\item{deparse}{(\code{logical}) flag specifying whether to return code as \code{character} (\code{deparse = TRUE}) or as
\code{expression} (\code{deparse = FALSE}).}

\item{datanames}{(\code{character}) vector of data set names to return the code for.}
}
\value{
Either string or expression representing code used to create the requested data sets.
}
\description{
Retrieve code from \code{teal_data} object.
}
\details{
Retrieve code stored in \verb{@code}, which (in principle) can be used to recreate all objects found in \verb{@env}.
Use \code{datanames} to limit the code to one or more of the data sets enumerated in \verb{@datanames}.
If the code has not passed verification, a warning will be prepended.
}
\examples{

tdata1 <- teal_data()
tdata1 <- within(tdata1, {
  a <- 1
  b <- a^5
  c <- list(x = 2)
})
get_code(tdata1)
datanames(tdata1) <- c("a", "b", "c")
get_code(tdata1, datanames = "a")
get_code(tdata1, datanames = "b")

tdata2 <- teal_data(x1 = iris, code = "x1 <- iris")
get_code(tdata2)
get_code(verify(tdata2))
}
