% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_key.R
\name{get_dataset_1}
\alias{get_dataset_1}
\alias{get_dataset_2}
\alias{get_keys.JoinKeySet}
\title{Getter for attributes in \code{JoinKeySet} object}
\usage{
get_dataset_1(join_key_set_object)

get_dataset_2(join_key_set_object)

\method{get_keys}{JoinKeySet}(join_key_set_object)
}
\arguments{
\item{join_key_set_object}{(\code{JoinKeySet}) object to retrieve attribute from.}
}
\value{
\code{dataset_1}, \code{dataset_2} or \code{key} as \code{character(1)}
}
\description{
Internal methods for \code{JoinKeySet} operations
}
\keyword{internal}
