% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_join_keys.R, R/join_keys.R
\name{get_join_keys}
\alias{get_join_keys}
\alias{get_join_keys<-}
\alias{join_keys<-.JoinKeys}
\alias{join_keys<-.teal_data}
\title{Function to get join keys from a `` object}
\usage{
get_join_keys(data)

get_join_keys(data, dataset_1, dataset_2 = NULL) <- value

\method{join_keys}{JoinKeys}(data, dataset_1, dataset_2 = NULL) <- value

\method{join_keys}{teal_data}(data, dataset_1, dataset_2 = NULL) <- value
}
\arguments{
\item{data}{`` - object to extract the join keys}

\item{dataset_1}{(\code{character}) one dataset name}

\item{dataset_2}{(\code{character}) other dataset name}

\item{value}{value to assign}
}
\value{
Either \code{JoinKeys} object or \code{NULL} if no join keys
}
\description{
Function to get join keys from a `` object
}
