% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_join_keys.R
\name{get_join_keys}
\alias{get_join_keys}
\alias{get_join_keys.teal_data}
\alias{get_join_keys.TealData}
\alias{get_join_keys.JoinKeys}
\alias{get_join_keys<-}
\alias{get_join_keys<-.JoinKeys}
\alias{get_join_keys<-.teal_data}
\alias{get_join_keys<-.TealData}
\title{Function to get join keys from a `` object}
\usage{
get_join_keys(data)

\method{get_join_keys}{teal_data}(data)

\method{get_join_keys}{TealData}(data)

\method{get_join_keys}{JoinKeys}(data)

get_join_keys(data, dataset_1, dataset_2 = NULL) <- value

\method{get_join_keys}{JoinKeys}(data, dataset_1, dataset_2 = NULL) <- value

\method{get_join_keys}{teal_data}(data, dataset_1, dataset_2 = NULL) <- value

\method{get_join_keys}{TealData}(data, dataset_1, dataset_2 = NULL) <- value
}
\arguments{
\item{data}{`` - object to extract the join keys}
}
\value{
Either \code{JoinKeys} object or \code{NULL} if no join keys
}
\description{
Function to get join keys from a `` object
}
