% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_join_keys.R
\name{get_join_keys}
\alias{get_join_keys}
\alias{get_join_keys<-}
\title{Function to get join keys from a `` object}
\usage{
get_join_keys(data)

get_join_keys(data, dataset_1, dataset_2 = NULL) <- value
}
\arguments{
\item{data}{`` - object to extract the join keys}

\item{dataset_1}{(\code{character(1)}) dataset name.}

\item{dataset_2}{(optional \code{character(1)}) other dataset name. In case it is omitted, then it
will create a primary key for \code{dataset_1}.}

\item{value}{value to assign}
}
\value{
Either \code{join_keys} object or \code{NULL} if no join keys
}
\description{
Function to get join keys from a `` object
}
