% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{get_join_keys}
\alias{get_join_keys}
\alias{get_join_keys<-}
\title{Deprecated function to retrieve \code{join_keys} from dataset.}
\usage{
get_join_keys(...)

get_join_keys(...) <- value
}
\arguments{
\item{...}{data argument.}

\item{value}{value to assign}
}
\value{
nothing
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}\cr
This function has been deprecated. Please see \code{\link[=join_keys]{join_keys()}} instead.
Find more information on \href{https://github.com/insightsengineering/teal/discussions/945}{Migration guide}.
}
