% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{get_join_keys}
\alias{get_join_keys}
\alias{get_join_keys<-}
\title{Function to get join keys from a `` object}
\usage{
get_join_keys(data)

get_join_keys(data, dataset_1, dataset_2 = NULL) <- value
}
\arguments{
\item{data}{`` - object to extract the join keys}

\item{dataset_1, dataset_2}{(\code{character(1)}) dataset names. If \code{dataset_2} is omitted,
a primary key for \code{dataset_1} is created.}

\item{value}{value to assign}
}
\value{
Either \code{join_keys} object or \code{NULL} if no join keys
}
\description{
Function to get join keys from a `` object
}
