% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_reproducible.R
\name{is_reproducible}
\alias{is_reproducible}
\title{Reproducibility Check for \verb{@code} Slot in \code{teal_data}}
\usage{
is_reproducible(teal_data)
}
\description{
Reproducibility Check for \verb{@code} Slot in \code{teal_data}
}
\examples{
tdata1 <- teal_data()
tdata1 <- within(tdata1, {
   a <- 1
   b <- a^5
   c <- list(x = 2)
})
is_reproducible(tdata1)
}
