% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_key.R
\name{join_key}
\alias{join_key}
\title{Create a relationship between a pair of datasets}
\usage{
join_key(dataset_1, dataset_2 = dataset_1, keys)
}
\arguments{
\item{dataset_1}{(\code{character}) one dataset name}

\item{dataset_2}{(optional \code{character}) other dataset name. In case it is omitted, then it
will create a primary key for \code{dataset_1}.}

\item{keys}{(optionally named \code{character}) where \code{names(keys)} are columns in \code{dataset_1}
with relationship to columns of \code{dataset_2} given by the elements in \code{keys}.
If \code{names(keys)} is \code{NULL} then the same column names are used for both \code{dataset_1}
and \code{dataset_2}.}
}
\value{
object of class \code{JoinKeySet} to be passed into \code{join_keys} function.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
\details{
\code{join_key()} will create a relationship for the variables on a pair
of datasets.
}
\seealso{
\code{\link[=join_keys]{join_keys()}}
}
