% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JoinKeys.R, R/join_keys.R
\name{join_keys}
\alias{join_keys}
\alias{cdisc_join_keys}
\alias{join_keys<-}
\alias{join_keys<-.Placeholder}
\alias{[.Placeholder}
\alias{[<-.Placeholder}
\title{Create a \code{JoinKeys} out of a list of \code{JoinKeySet} objects}
\usage{
join_keys(...)

cdisc_join_keys(...)

join_keys(join_keys_obj) <- value

\method{join_keys}{Placeholder}(join_keys_obj) <- value

\method{[}{Placeholder}(join_keys_obj, dataset_1 = NULL, dataset_2 = NULL)

\method{[}{Placeholder}(join_keys_obj, dataset_1, dataset_2 = dataset_1) <- value
}
\arguments{
\item{...}{optional, a \code{JoinKeySet} objects created using the \code{join_key} function.}

\item{join_keys_obj}{(\code{JoinKeys}) object to extract the join keys}

\item{value}{(\code{character} vector) value to assign.}

\item{dataset_1}{(\code{character}) name of first dataset.}

\item{dataset_2}{(\code{character}) name of second dataset.}
}
\value{
\code{JoinKeys}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
\details{
Note that join keys are symmetric although the relationship only needs
to be specified once.

\code{cdisc_join_keys} is a wrapper around \code{join_keys} that sets the default
join keys for CDISC datasets. It is used internally by \code{cdisc_data} to
set the default join keys for CDISC datasets.

The setter assignment \code{join_keys() <- ...} will only work for an empty
\code{JoinKey} object, otherwise \code{mutate_join_keys()} must be used.

The setter assignment \code{join_keys() <- ...} will only work for an empty
\code{JoinKey} object, otherwise \code{mutate_join_keys()} must be used.

Getter for JoinKeys that returns the relationship between pairs of datasets.

Setter via index directly (bypassing the need to use \code{join_key()}).
When \code{dataset_2} is omitted, it will create a primary key with \code{dataset_2 = dataset_1}.
}
\examples{
# setting join keys
join_keys(
  join_key("dataset_A", "dataset_B", c("col_1" = "col_a")),
  join_key("dataset_A", "dataset_C", c("col_2" = "col_x", "col_3" = "col_y"))
)
# or
jk <- join_keys()
jk["dataset_A", "dataset_B"] <- c("col_1" = "col_a")
jk["dataset_A", "dataset_C"] <- c("col_2" = "col_x", "col_3" = "col_y")


# default CDISC join keys
cdisc_join_keys(join_key("dataset_A", "dataset_B", c("col_1" = "col_a")), "ADTTE")

# Using the setter (assignment)
jk <- new_join_keys()
join_keys(jk)
join_keys(jk) <- join_key("ds1", "ds2", "some_col")
join_keys(jk) <- join_key("ds3", "ds4", "some_col2")
join_keys(jk)["ds1", "ds3"] <- "some_col3"
# Getter for JoinKeys
jk <- new_join_keys()
join_keys(jk) <- join_key("ds1", "ds2", "some_col")
jk["ds1", "ds2"]
jk["ds1"]
jk[["ds1"]]
# Setter via index
jk <- new_join_keys()
join_keys(jk) <- join_key("ds1", "ds2", "(original) pair key")
# overwrites previously defined key
jk["ds1", "ds2"] <- "(new) pair key"
# Creates primary key by only defining `dataset_1`
jk["ds1"] <- "primary_key"
jk
}
