% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdisc_join_keys.R, R/join_keys.R
\name{cdisc_join_keys}
\alias{cdisc_join_keys}
\alias{join_keys}
\alias{join_keys.join_keys}
\alias{join_keys.teal_data}
\alias{join_keys.TealData}
\alias{join_keys.default}
\alias{join_keys<-}
\alias{join_keys<-.join_keys}
\alias{c.join_keys}
\alias{join_keys<-.teal_data}
\alias{[.join_keys}
\alias{[<-.join_keys}
\alias{[[<-.join_keys}
\title{Create a \code{join_keys} out of a list of \code{join_key_set} objects}
\usage{
cdisc_join_keys(...)

join_keys(...)

\method{join_keys}{join_keys}(...)

\method{join_keys}{teal_data}(...)

\method{join_keys}{TealData}(...)

\method{join_keys}{default}(...)

join_keys(join_keys_obj) <- value

\method{join_keys}{join_keys}(join_keys_obj) <- value

\method{c}{join_keys}(...)

\method{join_keys}{teal_data}(join_keys_obj) <- value

\method{[}{join_keys}(join_keys_obj, dataset_1 = NULL, keep_all_foreign_keys = FALSE)

\method{[}{join_keys}(join_keys_obj, dataset_1) <- value

\method{[[}{join_keys}(join_keys_obj, dataset_1) <- value
}
\arguments{
\item{...}{(optional), when no argument is given the empty constructor is called.
Otherwise, when called with only one argument of type: \code{join_keys} or  \code{teal_data}
it will return the \code{join_keys} of that object.
When called with 1 or more \code{join_key_set} it will create a new \code{join_keys} with
constructed from the arguments.}

\item{join_keys_obj}{(\code{join_keys}) object to extract the join keys}

\item{value}{(\code{character} vector) value to assign.}

\item{dataset_1}{(\code{character}) name of first dataset.}
}
\value{
\code{join_keys} object.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
\details{
\code{cdisc_join_keys} is a wrapper around \code{join_keys} that sets the default
join keys for CDISC datasets. It is used internally by \code{cdisc_data} to
set the default join keys for CDISC datasets.

\itemize{
\item \code{join_keys()}: When called without arguments it will return an
empty constructor.
\item \code{join_keys(x)}: When called with a single argument it will return the \code{join_keys}
object contained in \code{x} (if it contains a \code{join_keys} object).
\item \code{join_keys(...)}: When called with a single or more \code{join_key_set} parameters it will
create a new object.
}

Note that join keys are created symmetrically, that is, if \code{dat1} and \code{dat2}
have a join key of \code{col1}, then 2 join keys are created, \verb{dat1 → dat2} and
\verb{dat2 → dat1}. The only exception is for a primary key.

The setter assignment \code{join_keys(obj) <- ...} will merge obj and \code{...} if obj
is not empty.

Getter for \code{join_keys} that returns the relationship between pairs of datasets.

Setter via index directly (bypassing the need to use \code{join_key()}).
When \code{dataset_2} is omitted, it will create a primary key with \code{dataset_2 = dataset_1}.
}
\examples{

# Default CDISC join keys

cdisc_join_keys(join_key("dataset_A", "dataset_B", c("col_1" = "col_a")), "ADTTE")
# Setting join keys ----

jk <- join_keys(
  join_key("dataset_A", "dataset_B", c("col_1" = "col_a")),
  join_key("dataset_A", "dataset_C", c("col_2" = "col_x", "col_3" = "col_y"))
)
jk

# or
jk <- join_keys()
jk[["dataset_A"]][["dataset_B"]] <- c("col_1" = "col_a")
jk[["dataset_A"]][["dataset_C"]] <- c("col_2" = "col_x", "col_3" = "col_y")
jk

td <- teal_data(join_keys = join_keys(join_key("a", "b", "c")))
join_keys(td)

jk <- join_keys()
join_keys(jk)

jk <- join_keys()
jk <- c(jk, join_keys(join_key("a", "b", "c")))
jk <- c(jk, join_keys(join_key("a", "b", "c"), join_key("a", "b2", "c")))

# Using the setter (assignment) ----

jk <- join_keys()
join_keys(jk) <- join_key("ds1", "ds2", "some_col")
join_keys(jk) <- c(join_keys(jk), join_key("ds3", "ds4", "some_col2"))

join_keys(jk)[["ds1"]][["ds3"]] <- "some_col3"
jk

c(join_keys(join_key("a", "b", "c")), join_keys(join_key("a", "d2", "c")))

# Setter for join_keys within teal_data ----

td <- teal_data()
join_keys(td)[["ds1"]][["ds2"]] <- "key1"
join_keys(td)[["ds2"]][["ds2"]] <- "key2"
join_keys(td) <- c(join_keys(td), join_keys(join_key("ds3", "ds2", "key3")))
join_keys(td)

# Getter for join_keys ----

jk <- join_keys()
jk[["ds1"]][["ds2"]] <- "some_col"
jk[["ds1"]][["ds3"]] <- "some_col2"

jk["ds1"]
jk[1:2]
jk[c("ds1", "ds2")]

# Setter via index ----

jk <- join_keys(
  join_key("ds1", "ds2", "col12"),
  join_key("ds3", "ds4", "col34")
)

# overwrites previously defined key
jk["ds1"] <- list(ds2 = "(new)co12")
jk["ds1"] <- list(ds3 = "col13", ds4 = "col14")
jk

jk[c("ds1", "ds2")] <- list(ds5 = "col*5")
jk[c(1, 2)] <- list(ds5 = "col**5")

# Creates primary key by only defining `dataset_1`
jk["ds1"] <- "primary_key"
jk

jk <- join_keys()
jk[["ds1"]] <- list()
jk[["ds2"]][["ds3"]] <- "key"

jk <- join_keys()
jk[["ds1"]] <- list()
jk[["ds2"]][["ds3"]] <- "key"
jk[["ds4"]] <- list(ds5 = "new")

jk <- join_keys()
jk[["ds2"]][["ds3"]] <- "key"
jk[["ds2"]][["ds3"]] <- NULL
jk
}
