% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_join_keys.R
\name{merge_join_keys}
\alias{merge_join_keys}
\title{Merging a list (or one) of \code{JoinKeys} objects into the current \code{JoinKeys} object}
\usage{
merge_join_keys(keys_1, keys_2)
}
\arguments{
\item{keys_1}{\code{JoinKeys} object}

\item{keys_2}{\code{list} of \code{JoinKeys} objects or single \code{JoinKeys} object}
}
\value{
(\code{JoinKeys}) a new object with the resulting merge
}
\description{
Merging a list (or one) of \code{JoinKeys} objects into the current \code{JoinKeys} object
}
\examples{
jk1 <- new_join_keys()
jk1["ds1", "ds2"] <- "some_col"
jk2 <- new_join_keys()
jk2["ds1", "ds3"] <- "new_col"
merge_join_keys(jk1, jk2)
}
