% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_keys.R
\name{merge_join_keys}
\alias{merge_join_keys}
\alias{merge_join_keys.default}
\alias{merge_join_keys.Placeholder}
\title{Merging a list (or one) of \code{JoinKeys} objects into the current \code{JoinKeys} object}
\usage{
merge_join_keys(join_keys_obj, new_join_keys)

\method{merge_join_keys}{default}(join_keys_obj, new_join_keys)

\method{merge_join_keys}{Placeholder}(join_keys_obj, new_join_keys)
}
\arguments{
\item{join_keys_obj}{(\code{JoinKeys}) object to merge the new_join_keys.}

\item{new_join_keys}{\code{list} of \code{JoinKeys} objects or single \code{JoinKeys} object}
}
\value{
a new \code{JoinKeys} object with the resulting merge.
}
\description{
Merging a list (or one) of \code{JoinKeys} objects into the current \code{JoinKeys} object
}
\examples{
jk1 <- new_join_keys()
jk1["ds1", "ds2"] <- "some_col"

jk2 <- new_join_keys()
jk2["ds1", "ds3"] <- "new_col"

merge_join_keys(jk1, jk2)
}
