% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_keys.R
\name{mutate_join_keys}
\alias{mutate_join_keys}
\alias{mutate_join_keys.JoinKeys}
\alias{mutate_join_keys.teal_data}
\title{Mutate \code{JoinKeys} with a new values}
\usage{
mutate_join_keys(x, dataset_1, dataset_2, value)

\method{mutate_join_keys}{JoinKeys}(x, dataset_1, dataset_2, value)

\method{mutate_join_keys}{teal_data}(x, dataset_1, dataset_2, value)
}
\arguments{
\item{x}{(\code{JoinKeys}) object to be modified}

\item{dataset_1}{(\code{character}) one dataset name}

\item{dataset_2}{(\code{character}) other dataset name}

\item{value}{(named \code{character}) column names used to join}
}
\value{
modified \code{JoinKeys} object
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
Mutate \code{JoinKeys} with a new values
}
\examples{

# JoinKeys ----

jk <- join_keys()
join_keys(jk) <- join_key("ds1", "ds2", "some_col")
mutate_join_keys(jk, "ds2", "ds3", "another")

# teal_data ----

ADSL <- teal.data::example_cdisc_data("ADSL")
ADRS <- teal.data::example_cdisc_data("ADRS")

x <- cdisc_data(
  "ADSL" = ADSL,
  "ADRS" = ADRS
)
join_keys(x)["ADSL", "ADRS"]

join_keys(x) <- mutate_join_keys(x, "ADSL", "ADRS", c("COLUMN1" = "COLUMN2"))
join_keys(x)["ADSL", "ADRS"]
}
