% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_keys-names.R
\name{names<-.join_keys}
\alias{names<-.join_keys}
\title{The Names of an \code{join_keys} Object}
\usage{
\method{names}{join_keys}(x) <- value
}
\arguments{
\item{x}{an \R object.}

\item{value}{a character vector of up to the same length as \code{x}, or
    \code{NULL}.}
}
\value{
For \code{names}, it returns \code{NULL} or a character vector containing the names of datasets.\cr
For \code{names<-.join_keys}, it returns the updated 'join_keys' object with the keys modified.
(Note that the value of names(x) <- value is that of the assignment,
value, not the return value from the left-hand side.)
}
\description{
The Names of an \code{join_keys} Object
}
