% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_keys-names.R
\name{names<-.join_keys}
\alias{names<-.join_keys}
\title{The names of a \code{join_keys} object}
\usage{
\method{names}{join_keys}(x) <- value
}
\arguments{
\item{x}{an \R object.}

\item{value}{a character vector of up to the same length as \code{x}, or
    \code{NULL}.}
}
\description{
The names of a \code{join_keys} object
}
