% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teal_data-class.R
\name{new_teal_data}
\alias{new_teal_data}
\title{Initialize \code{teal_data} object}
\usage{
new_teal_data(
  data,
  code = character(0),
  join_keys = join_keys(),
  datanames = names(data)
)
}
\arguments{
\item{data}{(\verb{named list}) of data objects.}

\item{code}{(\code{character} or \code{language}) code to reproduce the \code{data}.
Accepts and stores comments also.}

\item{join_keys}{(\code{join_keys}) object}

\item{datanames}{(\code{character}) names of datasets passed to \code{data}.
Needed when non-dataset objects are needed in the \code{env} slot.}
}
\description{
Initialize \code{teal_data} object
}
\keyword{internal}
