% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parents.R
\name{parent}
\alias{parent}
\title{Getter and setter for specific parent}
\usage{
parent(join_keys_obj, dataset_name)
}
\arguments{
\item{join_keys_obj}{(\code{JoinKeys}) object to retrieve.}

\item{dataset_name}{(\code{character(1)})}
}
\description{
Getter and setter for specific parent
}
\examples{
jk <- join_keys()
parent(jk, "ds1")
parents(jk) <- list("ds2" = "ds3")
parent(jk, "ds2")
}
