% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parents.R
\name{parents}
\alias{parents}
\alias{parents.Placeholder}
\alias{parents<-}
\alias{parents<-.Placeholder}
\alias{parents[}
\alias{parents[.Placeholder}
\alias{parents[[<-}
\alias{parents[<-}
\alias{parents[<-.Placeholder}
\title{A name}
\usage{
parents(join_keys_obj)

\method{parents}{Placeholder}(join_keys_obj)

parents(join_keys_obj) <- value

\method{parents}{Placeholder}(join_keys_obj) <- value

`parents[`(join_keys_obj, dataset)

\method{parents[}{Placeholder}(join_keys_obj, dataset)

`parents[[`(join_keys_obj, dataset) <- value

`parents[`(join_keys_obj, dataset) <- value

\method{parents[}{Placeholder}(join_keys_obj, dataset) <- value
}
\description{
A name
}
\examples{
jk <- new_join_keys()
parents(jk)
jk <- new_join_keys()
parents(jk) <- list(ADSL = "ADTTE")
jk <- new_join_keys()
parents(jk)["ADTTE"] <- "ADSL"
jk <- new_join_keys()
parents(jk)["ADTTE"] <- "ADSL"
parents(jk)[["YADA"]]
jk <- new_join_keys()
parents(jk)["ADTTE"] <- "ADSL"
}
