% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parents.R
\name{parents}
\alias{parents}
\alias{parents.join_keys}
\alias{parents<-}
\alias{parents<-.join_keys}
\title{Getter and setter functions for parents attribute of \code{join_keys}}
\usage{
parents(join_keys_obj)

\method{parents}{join_keys}(join_keys_obj)

parents(join_keys_obj) <- value

\method{parents}{join_keys}(join_keys_obj) <- value
}
\arguments{
\item{join_keys_obj}{(\code{join_keys}) object to retrieve or manipulate.}

\item{value}{(\code{list}) named list of character values}
}
\value{
a list of \code{character} representing the parents.
}
\description{
Getter and setter functions for parents attribute of \code{join_keys}
}
\examples{
jk <- join_keys()
parents(jk)
jk <- join_keys()
parents(jk) <- list(ADSL = "ADTTE")
jk <- join_keys()
parents(jk) <- list(ds1 = "ds2", "ds3" = "ds4")
parents(jk)["ds5"] <- "ds6"
parents(jk)["ds6"] <- "ds7"
}
