% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_keys-parents.R
\name{parents}
\alias{parents}
\alias{parents.join_keys}
\alias{parents.teal_data}
\alias{parents<-}
\alias{parents<-.join_keys}
\alias{parents<-.teal_data}
\title{Getter and setter functions for parents attribute of \code{join_keys}}
\usage{
parents(x)

\method{parents}{join_keys}(x)

\method{parents}{teal_data}(x)

parents(x) <- value

\method{parents}{join_keys}(x) <- value

\method{parents}{teal_data}(x) <- value
}
\arguments{
\item{x}{(\code{join_keys}) object to retrieve or manipulate.}

\item{value}{(\code{list}) named list of character values}
}
\value{
a list of \code{character} representing the parents.
}
\description{
Getter and setter functions for parents attribute of \code{join_keys}
}
\examples{
jk <- default_cdisc_join_keys["ADEX"]
parents(jk)

td <- cdisc_data(
  ADSL = teal.data::rADSL,
  ADTTE = teal.data::rADTTE
)
parents(td)

jk <- join_keys(
  join_key("ds1", "ds2", "id"),
  join_key("ds5", "ds6", "id"),
  join_key("ds7", "ds6", "id")
)
parents(jk) <- list()
parents(jk) <- list(ds1 = "ds2")
parents(jk)["ds5"] <- "ds6"
parents(jk)["ds6"] <- "ds7"

td <- cdisc_data(
  ADSL = teal.data::rADSL,
  ADTTE = teal.data::rADTTE,
  ADRS = teal.data::rADRS
)

parents(td) <- list("ADTTE" = "ADSL") # replace existing
parents(td)["ADRS"] <- "ADSL" # add new parent
}
