% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_keys-parents.R
\name{parent}
\alias{parent}
\alias{parents}
\alias{parents.join_keys}
\alias{parents.teal_data}
\alias{parents<-}
\alias{parents<-.join_keys}
\alias{parents<-.teal_data}
\title{Getter and setter functions for parents attribute of \code{join_keys}}
\usage{
parent(x, dataset_name)

parents(x)

\method{parents}{join_keys}(x)

\method{parents}{teal_data}(x)

parents(x) <- value

\method{parents}{join_keys}(x) <- value

\method{parents}{teal_data}(x) <- value
}
\arguments{
\item{x}{(\code{join_keys} or \code{teal_data}) object that contains "parents" information
to retrieve or manipulate.}

\item{dataset_name}{(\code{character(1)}) Name of dataset to query on their parent.}

\item{value}{(\code{list}) named list of character values}
}
\value{
For \code{parent(x, dataset_name)} returns \code{NULL} if parent does not exist.

a list of \code{character} representing the parents.
}
\description{
\code{parents()} facilitates the creation of dependencies between datasets by
assigning a parent-child relationship.

Each element is defined by a list element, where \code{list("child" = "parent")}.
}
\section{Methods (by class)}{
\itemize{
\item \code{parents(join_keys)}: Retrieves parents of \code{join_keys} object.

\item \code{parents(teal_data)}: Retrieves parents of \code{join_keys} inside \code{teal_data} object.

}}
\section{Functions}{
\itemize{
\item \code{parent()}: Getter for individual parent

\item \code{parents(x) <- value}: Assignment of parents in \code{join_keys} object.

\item \code{parents(join_keys) <- value}: Assignment of parents of \code{join_keys} object.

\item \code{parents(teal_data) <- value}: Assignment of parents of \code{join_keys} inside \code{teal_data} object.

}}
\examples{

# Get individual parent ---

parent(jk, "ds2")
parent(td, "ADTTE")
# Get parents of join_keys ---

jk <- default_cdisc_join_keys["ADEX"]
parents(jk)

# Get parents of join_keys inside teal_data object ---

td <- teal_data(
  ADSL = teal.data::rADSL,
  ADTTE = teal.data::rADTTE,
  ADRS = teal.data::rADRS,
  join_keys = default_cdisc_join_keys[c("ADSL", "ADTTE", "ADRS")]
)
parents(td)

# Assigment of parents ---

jk <- join_keys(
  join_key("ds1", "ds2", "id"),
  join_key("ds5", "ds6", "id"),
  join_key("ds7", "ds6", "id")
)

parents(jk) <- list(ds2 = "ds1")

# Setting individual parent-child relationship

parents(jk)["ds6"] <- "ds5"
parents(jk)["ds7"] <- "ds6"

# Assigment of parents of join_keys inside teal_data object ---

parents(td) <- list("ADTTE" = "ADSL") # replace existing
parents(td)["ADRS"] <- "ADSL" # add new parent
}
\seealso{
\code{\link[=join_keys]{join_keys()}}
}
