% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rADSL}
\alias{rADSL}
\title{Random patient listing dataset}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 400 rows and 55 columns.
}
\source{
\link[random.cdisc.data:cadsl]{random.cdisc.data::cadsl}
}
\usage{
rADSL
}
\description{
Random patient listing dataset
}
\keyword{datasets}
