% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rADVS}
\alias{rADVS}
\title{Random data \code{rADVS}}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 16800 rows and 87 columns.
}
\source{
internal
}
\usage{
rADVS
}
\description{
Random data \code{rADVS}.
}
\keyword{datasets}
\keyword{internal}
