% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_keys.R
\name{split_join_keys}
\alias{split_join_keys}
\alias{split_join_keys.default}
\alias{split_join_keys.Placeholder}
\title{Split the \code{JoinKeys} object into a named list of join keys objects with an element for each dataset}
\usage{
split_join_keys(join_keys_obj)

\method{split_join_keys}{default}(join_keys_obj)

\method{split_join_keys}{Placeholder}(join_keys_obj)
}
\value{
(\code{list}) a list of \code{JoinKeys} object
}
\description{
Split the \code{JoinKeys} object into a named list of join keys objects with an element for each dataset
}
\examples{
jk <- new_join_keys()
jk["ds1", "ds2"] <- "some_col"
jk["ds1", "ds3"] <- "new_col"
split_join_keys(jk)
}
