% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JoinKeys.R
\name{[.JoinKeys}
\alias{[.JoinKeys}
\alias{[<-.JoinKeys}
\title{Getter for JoinKeys that returns the relationship between pairs of datasets}
\usage{
\method{[}{JoinKeys}(x, dataset_1, dataset_2 = dataset_1)

\method{[}{JoinKeys}(x, dataset_1, dataset_2 = dataset_1) <- value
}
\arguments{
\item{x}{JoinKeys object to extract the join keys}

\item{dataset_1}{(\code{character}) name of first dataset.}

\item{dataset_2}{(\code{character}) name of second dataset.}

\item{value}{value to assign}
}
\description{
Getter for JoinKeys that returns the relationship between pairs of datasets
}
\keyword{internal}
