% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teal.data.R
\docType{package}
\name{teal.data-package}
\alias{teal.data}
\alias{teal.data-package}
\title{teal.data: Reproducible data model for \code{teal} applications}
\description{
This package extends \code{teal} applications by module which
stores the data with they relationships (keys) and reproducible
code. Package offers also to load data from files, databases and
it's easy extendable by another data sources.
}
\author{
\strong{Maintainer}: NEST \email{basel.nestcicd@roche.com}

Authors:
\itemize{
  \item Nikolas Burkoff \email{nikolas.burkoff@roche.com}
  \item Mahmoud Hallal \email{mahmoud.hallal@roche.com}
  \item Dawid Kaledkowski \email{dawid.kaledkowski@roche.com}
  \item Maciej Nasinski \email{maciej.nasinski@contractors.roche.com}
  \item Konrad Pagacz \email{konrad.pagacz@contractors.roche.com}
  \item Pawel Rucki \email{pawel.rucki@contractors.roche.com}
  \item Junlue Zhao \email{zhaoj88@gene.com}
}

Other contributors:
\itemize{
  \item Tadeusz Lewandowski \email{tadeusz.lewandowski@roche.com} [contributor]
  \item Roche [copyright holder, funder]
}

}
\keyword{internal}
