% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teal.data.R
\docType{package}
\name{teal.data-package}
\alias{teal.data}
\alias{teal.data-package}
\title{teal.data: Reproducible data model for \code{teal} applications}
\description{
This package extends \code{teal} applications by module which
stores the data with they relationships (keys) and reproducible
code. Package offers also to load data from files, databases and
it's easily extendable by another data sources.
}
\author{
\strong{Maintainer}: Dawid Kaledkowski \email{dawid.kaledkowski@roche.com}

Authors:
\itemize{
  \item Pawel Rucki \email{pawel.rucki@roche.com}
}

Other contributors:
\itemize{
  \item F. Hoffmann-La Roche AG [copyright holder, funder]
}

}
\keyword{internal}
