% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teal.data.R
\docType{package}
\name{teal.data-package}
\alias{teal.data}
\alias{teal.data-package}
\title{\code{teal.data}: Reproducible data model for \code{teal} applications}
\description{
This package extends \code{teal} applications by module which
stores the data with they relationships (keys) and reproducible
code. Package offers also to load data from files, databases and
it's easily extendable by another data sources.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://insightsengineering.github.io/teal.data/}
  \item \url{https://github.com/insightsengineering/teal.data/}
  \item Report bugs at \url{https://github.com/insightsengineering/teal.data/issues}
}

}
\author{
\strong{Maintainer}: Dawid Kaledkowski \email{dawid.kaledkowski@roche.com} (\href{https://orcid.org/0000-0001-9533-457X}{ORCID})

Authors:
\itemize{
  \item Aleksander Chlebowski \email{aleksander.chlebowski@contractors.roche.com}
  \item Marcin Kosinski \email{marcin.kosinski.mk1@roche.com}
  \item Andre Verissimo \email{andre.verissimo@roche.com}
  \item Pawel Rucki \email{pawel.rucki@roche.com}
  \item Mahmoud Hallal \email{mahmoud.hallal@roche.com}
  \item Nikolas Burkoff
  \item Maciej Nasinski
  \item Konrad Pagacz
  \item Junlue Zhao
}

Other contributors:
\itemize{
  \item Chendi Liao \email{chendi.liao@roche.com} [reviewer]
  \item Dony Unardi \email{unardid@gene.com} [reviewer]
  \item F. Hoffmann-La Roche AG [copyright holder, funder]
}

}
\keyword{internal}
