% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teal_data-class.R
\docType{class}
\name{teal_data-class}
\alias{teal_data-class}
\title{Reproducible data class.}
\description{
Reproducible data class basing on \linkS4class{qenv}.
}
\section{Slots}{

\describe{
\item{\code{code}}{(\code{expression}) to reproduce the environment}

\item{\code{env}}{(\code{environment}) environment which content was generated by the evaluation
of the \code{code} slot.}

\item{\code{id}}{(\code{integer}) random identifier of the code element to make sure uniqueness
when joining.}

\item{\code{warnings}}{(\code{character}) the warnings output when evaluating the code}

\item{\code{messages}}{(\code{character}) the messages output when evaluating the code}

\item{\code{join_keys}}{(\code{JoinKeys}) object}

\item{\code{datanames}}{(\code{character}) names of datasets in \code{env}. Needed when non-dataset
objects are needed in the \code{env} slot.}
}}

\keyword{internal}
