% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teal_data-class.R
\docType{class}
\name{teal_data-class}
\alias{teal_data-class}
\title{Reproducible data class.}
\description{
Reproducible data class basing on \code{\link[teal.code:qenv-class]{teal.code::qenv}}.
Don't interact with slots directly, it is suggested to use methods instead
(see \code{methods(class = "teal_data")}).
}
\section{Slots}{

\describe{
\item{\code{code}}{(\code{character}) representing code necessary to reproduce the environment.
To extract the \code{code} please use \code{\link[=get_code]{get_code()}}.}

\item{\code{env}}{(\code{environment}) environment which content was generated by the evaluation
of the \code{code} slot. To extract variables from the environment please use \code{\link[=get_var]{get_var()}} or [\code{[[}].}

\item{\code{id}}{(\code{integer}) random identifier of the code element to make sure uniqueness
when joining.}

\item{\code{warnings}}{(\code{character}) the warnings output when evaluating the code.
To extract the \code{warnings} use \code{\link[=get_warnings]{get_warnings()}}.}

\item{\code{messages}}{(\code{character}) the messages output when evaluating the code}

\item{\code{join_keys}}{(\code{JoinKeys}) object.
To extract the \code{join_keys} use \code{\link[=get_join_keys]{get_join_keys()}}.}

\item{\code{datanames}}{(\code{character}) names of datasets in \code{env}. Needed when non-dataset
objects are needed in the \code{env} slot.
To extract the \code{datanames} use \code{\link[=get_dataname]{get_dataname()}}.}
}}

\keyword{internal}
