% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teal_data.R
\name{teal_data}
\alias{teal_data}
\title{Teal Data}
\usage{
teal_data(
  ...,
  join_keys = teal.data::join_keys(),
  code = character(0),
  check = FALSE
)
}
\arguments{
\item{...}{any number of objects (presumably data objects) provided as \code{name = value} pairs.}

\item{join_keys}{(\code{join_keys}) or a single (\code{join_key_set})\cr
(optional) object with dataset column relationships used for joining.
If empty then no joins between pairs of objects}

\item{code}{(\code{character}, \code{language}) code to reproduce the datasets.}

\item{check}{(\code{logical}) reproducibility check - whether to perform a check that the pre-processing
code included in the object definitions actually produces those objects.
If \code{check} is true and preprocessing code is empty an error will be thrown.}
}
\value{
A \code{teal_data} object.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
Universal function to pass data to teal application.
}
\examples{
teal_data(x1 = iris, x2 = mtcars)

}
