% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_keys.R
\name{update_keys_given_parents}
\alias{update_keys_given_parents}
\title{Updates the keys of the datasets based on the parents.}
\usage{
update_keys_given_parents(join_keys_obj)
}
\value{
(\code{self}) invisibly for chaining
}
\description{
Updates the keys of the datasets based on the parents.
}
