% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parents.R
\name{update_keys_given_parents}
\alias{update_keys_given_parents}
\title{Updates the keys of the datasets based on the parents.}
\usage{
update_keys_given_parents(join_keys_obj)
}
\arguments{
\item{join_keys_obj}{(\code{join_keys}) object to update the keys.}
}
\value{
(\code{self}) invisibly for chaining
}
\description{
Updates the keys of the datasets based on the parents.
}
