% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify.R
\name{verify}
\alias{verify}
\alias{verify,teal_data-method}
\alias{verify,qenv.error-method}
\title{Reproducibility Check for \verb{@code} Slot in \code{teal_data}}
\usage{
verify(x)
}
\arguments{
\item{x}{\code{teal_data} object}
}
\value{
If the objects returned by the \verb{@code} slot in the input \code{teal_data} object exactly match
the objects in \code{teal_data@env}, the function updates the \verb{@verified} slot to \code{TRUE} in the \code{teal_data} object.
If the \verb{@code} fails to recreate objects in \code{teal_data@env}, an error is thrown.
}
\description{
Verifies the reproducibility of the \code{teal_data@env} environment using the \verb{@code} slot.
}
\examples{
tdata1 <- teal_data()
tdata1 <- within(tdata1, {
  a <- 1
  b <- a^5
  c <- list(x = 2)
})
verify(tdata1)

tdata2 <- teal_data(x1 = iris, code = "x1 = iris")
verify(tdata2)
verify(tdata2)@verified
tdata2@verified

\dontrun{
tdata3 <- teal_data()
tdata3 <- within(tdata3, {
  stop("error")
})
}
}
